refactor(cli): split moraine command modules
