+++ apps/moraine/src/cli.rs
@@ -1,0 +1,187 @@
+use clap::{Args, Parser, Subcommand, ValueEnum};
+use std::path::PathBuf;
+
+use crate::service::Service;
+
+#[derive(Debug, Clone, Copy, ValueEnum)]
+pub(crate) enum OutputFormat {
+    Auto,
+    Rich,
+    Plain,
+    Json,
+}
+
+#[derive(Debug, Parser)]
+#[command(
+    name = "moraine",
+    about = "Unified runtime control plane for Moraine services",
+    version = env!("CARGO_PKG_VERSION")
+)]
+pub(crate) struct Cli {
+    #[arg(long, global = true, value_name = "PATH")]
+    pub(crate) config: Option<PathBuf>,
+    #[arg(long, global = true, value_enum, default_value_t = OutputFormat::Auto)]
+    pub(crate) output: OutputFormat,
+    #[arg(long, global = true, default_value_t = false)]
+    pub(crate) verbose: bool,
+    #[command(subcommand)]
+    pub(crate) command: CliCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum CliCommand {
+    Up(UpArgs),
+    Down,
+    Status,
+    Logs(LogsArgs),
+    Db(DbArgs),
+    Clickhouse(ClickhouseArgs),
+    Config(ConfigArgs),
+    Run(RunArgs),
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct UpArgs {
+    #[arg(long)]
+    pub(crate) no_ingest: bool,
+    #[arg(long)]
+    pub(crate) monitor: bool,
+    #[arg(long)]
+    pub(crate) mcp: bool,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct LogsArgs {
+    #[arg(value_enum)]
+    pub(crate) service: Option<Service>,
+    #[arg(long, default_value_t = 200)]
+    pub(crate) lines: usize,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct DbArgs {
+    #[command(subcommand)]
+    pub(crate) command: DbCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum DbCommand {
+    Migrate,
+    Doctor,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct ClickhouseArgs {
+    #[command(subcommand)]
+    pub(crate) command: ClickhouseCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum ClickhouseCommand {
+    Install(ClickhouseInstallArgs),
+    Status,
+    Uninstall,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct ClickhouseInstallArgs {
+    #[arg(long)]
+    pub(crate) force: bool,
+    #[arg(long)]
+    pub(crate) version: Option<String>,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct ConfigArgs {
+    #[command(subcommand)]
+    pub(crate) command: ConfigCommand,
+}
+
+#[derive(Debug, Subcommand)]
+pub(crate) enum ConfigCommand {
+    Get(ConfigGetArgs),
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct ConfigGetArgs {
+    #[arg(value_name = "KEY")]
+    pub(crate) key: String,
+}
+
+#[derive(Debug, Args)]
+pub(crate) struct RunArgs {
+    #[arg(value_enum)]
+    pub(crate) service: Service,
+    #[arg(
+        trailing_var_arg = true,
+        allow_hyphen_values = true,
+        num_args = 0..
+    )]
+    pub(crate) args: Vec<String>,
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn clap_parses_clickhouse_install_flags() {
+        let cli = Cli::parse_from([
+            "moraine",
+            "clickhouse",
+            "install",
+            "--version",
+            "v25.12.5.44-stable",
+            "--force",
+        ]);
+        match cli.command {
+            CliCommand::Clickhouse(ClickhouseArgs {
+                command: ClickhouseCommand::Install(install),
+            }) => {
+                assert!(install.force);
+                assert_eq!(install.version.as_deref(), Some("v25.12.5.44-stable"));
+            }
+            _ => panic!("expected clickhouse install command"),
+        }
+    }
+
+    #[test]
+    fn clap_parses_config_get_key() {
+        let cli = Cli::parse_from(["moraine", "config", "get", "clickhouse.url"]);
+        match cli.command {
+            CliCommand::Config(ConfigArgs {
+                command: ConfigCommand::Get(get),
+            }) => assert_eq!(get.key, "clickhouse.url"),
+            _ => panic!("expected config get command"),
+        }
+    }
+
+    #[test]
+    fn clap_parses_run_passthrough_args() {
+        let cli = Cli::parse_from([
+            "moraine",
+            "--output",
+            "plain",
+            "run",
+            "mcp",
+            "--",
+            "--stdio",
+            "--transport",
+            "jsonrpc",
+        ]);
+        match cli.command {
+            CliCommand::Run(run) => {
+                assert_eq!(run.service, Service::Mcp);
+                assert_eq!(
+                    run.args,
+                    vec![
+                        "--stdio".to_string(),
+                        "--transport".to_string(),
+                        "jsonrpc".to_string(),
+                    ]
+                );
+            }
+            _ => panic!("expected run command"),
+        }
+    }
+}
