+++ apps/moraine/src/main.rs
@@ -1,3078 +1,21 @@
-use anyhow::{anyhow, bail, Context, Result};
-use clap::{Args, Parser, Subcommand, ValueEnum};
-use moraine_clickhouse::{ClickHouseClient, DoctorReport};
-use moraine_config::AppConfig;
-use ratatui::buffer::Buffer;
-use ratatui::layout::{Constraint, Rect};
-use ratatui::style::{Color, Modifier, Style};
-use ratatui::text::Line;
-use ratatui::widgets::{Block, BorderType, Borders, Cell, Paragraph, Row, Table, Widget, Wrap};
-use reqwest::{Client, Url};
-use serde::Deserialize;
-use sha2::{Digest, Sha256};
-use std::fs::{self, OpenOptions};
-use std::io::{IsTerminal, Read, Seek, SeekFrom, Write};
-use std::path::{Path, PathBuf};
-use std::process::{Command, ExitCode, Stdio};
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-use tokio::time::{sleep, Instant};
-
-#[cfg(unix)]
-use std::os::unix::fs::{symlink, PermissionsExt};
-
-const CLICKHOUSE_TEMPLATE: &str = include_str!("../../../config/clickhouse.xml");
-const USERS_TEMPLATE: &str = include_str!("../../../config/users.xml");
-
-const DEFAULT_CLICKHOUSE_TAG: &str = "v25.12.5.44-stable";
-const CH_URL_MACOS_X86_64: &str = "https://github.com/ClickHouse/ClickHouse/releases/download/v25.12.5.44-stable/clickhouse-macos";
-const CH_SHA_MACOS_X86_64: &str =
-    "8035b4b7905147156192216cc6937a29d0cd2775d481b5f297cdc11058cb68c4";
-const CH_URL_MACOS_AARCH64: &str = "https://github.com/ClickHouse/ClickHouse/releases/download/v25.12.5.44-stable/clickhouse-macos-aarch64";
-const CH_SHA_MACOS_AARCH64: &str =
-    "1a0edc37c6e5aa6c06a7cb00c8f8edd83a0df02f643e29185a8b3934eb860ac4";
-const CH_URL_LINUX_X86_64: &str = "https://github.com/ClickHouse/ClickHouse/releases/download/v25.12.5.44-stable/clickhouse-common-static-25.12.5.44-amd64.tgz";
-const CH_SHA_LINUX_X86_64: &str =
-    "3756d8b061f97abd79621df1a586f6ba777e8787696f21d82bc488ce5dbca2d7";
-const CH_URL_LINUX_AARCH64: &str = "https://github.com/ClickHouse/ClickHouse/releases/download/v25.12.5.44-stable/clickhouse-common-static-25.12.5.44-arm64.tgz";
-const CH_SHA_LINUX_AARCH64: &str =
-    "3d227e50109b0dab330ee2230f46d76f0360f1a61956443c37de5b7651fb488b";
-
-#[derive(Debug, Clone, Copy, ValueEnum)]
-enum OutputFormat {
-    Auto,
-    Rich,
-    Plain,
-    Json,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum OutputMode {
-    Rich,
-    Plain,
-    Json,
-}
-
-#[derive(Debug, Parser)]
-#[command(
-    name = "moraine",
-    about = "Unified runtime control plane for Moraine services",
-    version = env!("CARGO_PKG_VERSION")
-)]
-struct Cli {
-    #[arg(long, global = true, value_name = "PATH")]
-    config: Option<PathBuf>,
-    #[arg(long, global = true, value_enum, default_value_t = OutputFormat::Auto)]
-    output: OutputFormat,
-    #[arg(long, global = true, default_value_t = false)]
-    verbose: bool,
-    #[command(subcommand)]
-    command: CliCommand,
-}
-
-#[derive(Debug, Subcommand)]
-enum CliCommand {
-    Up(UpArgs),
-    Down,
-    Status,
-    Logs(LogsArgs),
-    Db(DbArgs),
-    Clickhouse(ClickhouseArgs),
-    Config(ConfigArgs),
-    Run(RunArgs),
-}
-
-#[derive(Debug, Args)]
-struct UpArgs {
-    #[arg(long)]
-    no_ingest: bool,
-    #[arg(long)]
-    monitor: bool,
-    #[arg(long)]
-    mcp: bool,
-}
-
-#[derive(Debug, Args)]
-struct LogsArgs {
-    #[arg(value_enum)]
-    service: Option<Service>,
-    #[arg(long, default_value_t = 200)]
-    lines: usize,
-}
-
-#[derive(Debug, Args)]
-struct DbArgs {
-    #[command(subcommand)]
-    command: DbCommand,
-}
-
-#[derive(Debug, Subcommand)]
-enum DbCommand {
-    Migrate,
-    Doctor,
-}
-
-#[derive(Debug, Args)]
-struct ClickhouseArgs {
-    #[command(subcommand)]
-    command: ClickhouseCommand,
-}
-
-#[derive(Debug, Subcommand)]
-enum ClickhouseCommand {
-    Install(ClickhouseInstallArgs),
-    Status,
-    Uninstall,
-}
-
-#[derive(Debug, Args)]
-struct ClickhouseInstallArgs {
-    #[arg(long)]
-    force: bool,
-    #[arg(long)]
-    version: Option<String>,
-}
-
-#[derive(Debug, Args)]
-struct ConfigArgs {
-    #[command(subcommand)]
-    command: ConfigCommand,
-}
-
-#[derive(Debug, Subcommand)]
-enum ConfigCommand {
-    Get(ConfigGetArgs),
-}
-
-#[derive(Debug, Args)]
-struct ConfigGetArgs {
-    #[arg(value_name = "KEY")]
-    key: String,
-}
-
-#[derive(Debug, Args)]
-struct RunArgs {
-    #[arg(value_enum)]
-    service: Service,
-    #[arg(
-        trailing_var_arg = true,
-        allow_hyphen_values = true,
-        num_args = 0..
-    )]
-    args: Vec<String>,
-}
-
-#[derive(Clone)]
-struct RuntimePaths {
-    root: PathBuf,
-    logs_dir: PathBuf,
-    pids_dir: PathBuf,
-    clickhouse_root: PathBuf,
-    clickhouse_config: PathBuf,
-    clickhouse_users: PathBuf,
-    service_bin_dir: PathBuf,
-    managed_clickhouse_dir: PathBuf,
-}
-
-#[derive(Clone, Copy, PartialEq, Eq, Debug, ValueEnum, serde::Serialize)]
-#[serde(rename_all = "lowercase")]
-enum Service {
-    #[value(name = "clickhouse")]
-    ClickHouse,
-    #[value(name = "ingest")]
-    Ingest,
-    #[value(name = "monitor")]
-    Monitor,
-    #[value(name = "mcp")]
-    Mcp,
-}
-
-impl Service {
-    fn name(self) -> &'static str {
-        match self {
-            Self::ClickHouse => "clickhouse",
-            Self::Ingest => "ingest",
-            Self::Monitor => "monitor",
-            Self::Mcp => "mcp",
-        }
-    }
-
-    fn pid_file(self) -> &'static str {
-        match self {
-            Self::ClickHouse => "clickhouse.pid",
-            Self::Ingest => "ingest.pid",
-            Self::Monitor => "monitor.pid",
-            Self::Mcp => "mcp.pid",
-        }
-    }
-
-    fn log_file(self) -> &'static str {
-        match self {
-            Self::ClickHouse => "clickhouse.log",
-            Self::Ingest => "ingest.log",
-            Self::Monitor => "monitor.log",
-            Self::Mcp => "mcp.log",
-        }
-    }
-
-    fn binary_name(self) -> Option<&'static str> {
-        match self {
-            Self::ClickHouse => None,
-            Self::Ingest => Some("moraine-ingest"),
-            Self::Monitor => Some("moraine-monitor"),
-            Self::Mcp => Some("moraine-mcp"),
-        }
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct HeartbeatRow {
-    latest: String,
-    queue_depth: u64,
-    files_active: u64,
-    #[serde(default)]
-    watcher_backend: String,
-    #[serde(default)]
-    watcher_error_count: u64,
-    #[serde(default)]
-    watcher_reset_count: u64,
-    #[serde(default)]
-    watcher_last_reset_unix_ms: u64,
-}
-
-#[derive(Debug, Deserialize)]
-struct LegacyHeartbeatRow {
-    latest: String,
-    queue_depth: u64,
-    files_active: u64,
-}
-
-#[derive(Clone, Copy, Debug)]
-struct ClickHouseAsset {
-    url: &'static str,
-    sha256: &'static str,
-    is_archive: bool,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct ServiceRuntimeStatus {
-    service: Service,
-    pid: Option<u32>,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-#[serde(tag = "state", rename_all = "snake_case")]
-enum HeartbeatSnapshot {
-    Available {
-        latest: String,
-        queue_depth: u64,
-        files_active: u64,
-        watcher_backend: String,
-        watcher_error_count: u64,
-        watcher_reset_count: u64,
-        watcher_last_reset_unix_ms: u64,
-    },
-    Unavailable,
-    Error {
-        message: String,
-    },
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct StatusSnapshot {
-    services: Vec<ServiceRuntimeStatus>,
-    monitor_url: Option<String>,
-    managed_clickhouse_installed: bool,
-    managed_clickhouse_path: String,
-    managed_clickhouse_version: Option<String>,
-    clickhouse_active_source: String,
-    clickhouse_active_source_path: Option<String>,
-    managed_clickhouse_checksum: String,
-    clickhouse_health_url: String,
-    status_notes: Vec<String>,
-    doctor: DoctorReport,
-    heartbeat: HeartbeatSnapshot,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct MigrationOutcome {
-    applied: Vec<String>,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct ServiceLogSection {
-    service: Service,
-    path: String,
-    exists: bool,
-    lines: Vec<String>,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct LogsSnapshot {
-    requested_lines: usize,
-    sections: Vec<ServiceLogSection>,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct ClickhouseStatusSnapshot {
-    managed_root: String,
-    clickhouse_exists: bool,
-    clickhouse_server_exists: bool,
-    clickhouse_client_exists: bool,
-    expected_version: String,
-    active_source: String,
-    active_source_path: Option<String>,
-    checksum_state: String,
-    installed_version: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, serde::Serialize)]
-#[serde(rename_all = "snake_case")]
-enum StartState {
-    Started,
-    AlreadyRunning,
-    /// A service endpoint is already healthy but not launcher-tracked, such as
-    /// an external ClickHouse endpoint or a live central MCP server whose PID
-    /// file was lost. Nothing was spawned.
-    AlreadyServing,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct StartOutcome {
-    service: Service,
-    state: StartState,
-    /// `None` when the service is alive but not launcher-tracked
-    /// (`StartState::AlreadyServing`).
-    pid: Option<u32>,
-    log_path: Option<String>,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct UpSnapshot {
-    clickhouse: StartOutcome,
-    migrations: MigrationOutcome,
-    services: Vec<StartOutcome>,
-    status: StatusSnapshot,
-}
-
-#[derive(Debug, Clone, serde::Serialize)]
-struct DownSnapshot {
-    stopped: Vec<Service>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    warning: Option<String>,
-}
-
-struct CliOutput {
-    mode: OutputMode,
-    verbose: bool,
-    unicode: bool,
-    width: u16,
-}
-
-impl CliOutput {
-    fn from_cli(cli: &Cli) -> Self {
-        let mode = match cli.output {
-            OutputFormat::Auto => {
-                if std::io::stdout().is_terminal() {
-                    OutputMode::Rich
-                } else {
-                    OutputMode::Plain
-                }
-            }
-            OutputFormat::Rich => OutputMode::Rich,
-            OutputFormat::Plain => OutputMode::Plain,
-            OutputFormat::Json => OutputMode::Json,
-        };
-        let unicode = std::env::var("LC_ALL")
-            .ok()
-            .or_else(|| std::env::var("LANG").ok())
-            .map(|v| !v.to_ascii_uppercase().contains("C"))
-            .unwrap_or(true);
-        let width = std::env::var("COLUMNS")
-            .ok()
-            .and_then(|v| v.parse::<u16>().ok())
-            .map(|v| v.clamp(72, 140))
-            .unwrap_or(100);
-
-        Self {
-            mode,
-            verbose: cli.verbose,
-            unicode,
-            width,
-        }
-    }
-
-    fn is_json(&self) -> bool {
-        self.mode == OutputMode::Json
-    }
-
-    fn section(&self, title: &str, lines: &[String]) {
-        match self.mode {
-            OutputMode::Plain => {
-                println!("{title}");
-                for line in lines {
-                    println!("  {line}");
-                }
-            }
-            OutputMode::Rich => {
-                let panel = render_panel(title, lines, self.width, self.unicode);
-                println!("{panel}");
-            }
-            OutputMode::Json => {}
-        }
-    }
-
-    fn table(&self, title: &str, headers: &[&str], rows: &[Vec<String>]) {
-        match self.mode {
-            OutputMode::Plain => print_plain_table(title, headers, rows),
-            OutputMode::Rich => {
-                let table = render_table(title, headers, rows, self.width, self.unicode);
-                println!("{table}");
-            }
-            OutputMode::Json => {}
-        }
-    }
-
-    fn line(&self, text: &str) {
-        if self.mode != OutputMode::Json {
-            println!("{text}");
-        }
-    }
-}
-
-fn render_panel(title: &str, lines: &[String], width: u16, unicode: bool) -> String {
-    let area = Rect::new(0, 0, width, (lines.len().max(1) as u16).saturating_add(2));
-    let mut buffer = Buffer::empty(area);
-    let mut block = Block::default()
-        .title(Line::from(title.to_string()))
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .border_style(Style::default().fg(Color::Cyan));
-    if !unicode {
-        block = block.border_set(ratatui::symbols::border::PLAIN);
-    }
-    let paragraph = Paragraph::new(lines.join("\n"))
-        .block(block)
-        .wrap(Wrap { trim: false })
-        .style(Style::default().fg(Color::White));
-    paragraph.render(area, &mut buffer);
-    buffer_to_string(&buffer)
-}
-
-fn render_table(
-    title: &str,
-    headers: &[&str],
-    rows: &[Vec<String>],
-    width: u16,
-    unicode: bool,
-) -> String {
-    let area = Rect::new(
-        0,
-        0,
-        width,
-        (rows.len().saturating_add(1) as u16).saturating_add(2),
-    );
-    let mut buffer = Buffer::empty(area);
-    let mut block = Block::default()
-        .title(Line::from(title.to_string()))
-        .borders(Borders::ALL)
-        .border_type(BorderType::Rounded)
-        .border_style(Style::default().fg(Color::Cyan));
-    if !unicode {
-        block = block.border_set(ratatui::symbols::border::PLAIN);
-    }
-
-    let header = Row::new(
-        headers
-            .iter()
-            .map(|h| Cell::from((*h).to_string()).style(Style::default().fg(Color::Yellow))),
-    )
-    .style(Style::default().add_modifier(Modifier::BOLD));
-    let data_rows = rows.iter().map(|row| Row::new(row.clone()));
-    let widths = headers
-        .iter()
-        .map(|_| Constraint::Percentage((100 / headers.len().max(1)) as u16))
-        .collect::<Vec<_>>();
-    let table = Table::new(data_rows, widths).header(header).block(block);
-    table.render(area, &mut buffer);
-    buffer_to_string(&buffer)
-}
-
-fn buffer_to_string(buffer: &Buffer) -> String {
-    let mut lines = Vec::new();
-    for y in 0..buffer.area.height {
-        let mut line = String::new();
-        for x in 0..buffer.area.width {
-            line.push_str(buffer[(x, y)].symbol());
-        }
-        while line.ends_with(' ') {
-            line.pop();
-        }
-        lines.push(line);
-    }
-    while lines.last().is_some_and(|line| line.is_empty()) {
-        lines.pop();
-    }
-    lines.join("\n")
-}
-
-fn print_plain_table(title: &str, headers: &[&str], rows: &[Vec<String>]) {
-    println!("{title}");
-    println!("{}", headers.join(" | "));
-    let divider = headers.iter().map(|_| "---").collect::<Vec<_>>().join("+");
-    println!("{divider}");
-    for row in rows {
-        println!("{}", row.join(" | "));
-    }
-}
-
-fn runtime_paths(cfg: &AppConfig) -> RuntimePaths {
-    let root = PathBuf::from(&cfg.runtime.root_dir);
-    let clickhouse_root = root.join("clickhouse");
-
-    RuntimePaths {
-        root,
-        logs_dir: PathBuf::from(&cfg.runtime.logs_dir),
-        pids_dir: PathBuf::from(&cfg.runtime.pids_dir),
-        clickhouse_config: clickhouse_root.join("config.xml"),
-        clickhouse_users: clickhouse_root.join("users.xml"),
-        clickhouse_root,
-        service_bin_dir: PathBuf::from(&cfg.runtime.service_bin_dir),
-        managed_clickhouse_dir: PathBuf::from(&cfg.runtime.managed_clickhouse_dir),
-    }
-}
-
-fn ensure_runtime_dirs(paths: &RuntimePaths) -> Result<()> {
-    fs::create_dir_all(&paths.root)
-        .with_context(|| format!("failed to create {}", paths.root.display()))?;
-    fs::create_dir_all(&paths.logs_dir)
-        .with_context(|| format!("failed to create {}", paths.logs_dir.display()))?;
-    fs::create_dir_all(&paths.pids_dir)
-        .with_context(|| format!("failed to create {}", paths.pids_dir.display()))?;
-
-    fs::create_dir_all(paths.clickhouse_root.join("data"))?;
-    fs::create_dir_all(paths.clickhouse_root.join("tmp"))?;
-    fs::create_dir_all(paths.clickhouse_root.join("log"))?;
-    fs::create_dir_all(paths.clickhouse_root.join("user_files"))?;
-    fs::create_dir_all(paths.clickhouse_root.join("format_schemas"))?;
-
-    Ok(())
-}
-
-fn pid_path(paths: &RuntimePaths, service: Service) -> PathBuf {
-    paths.pids_dir.join(service.pid_file())
-}
-
-fn clickhouse_internal_log_path(paths: &RuntimePaths) -> PathBuf {
-    paths
-        .clickhouse_root
-        .join("log")
-        .join("clickhouse-server.log")
-}
-
-fn legacy_clickhouse_pipe_log_path(paths: &RuntimePaths) -> PathBuf {
-    paths.logs_dir.join(Service::ClickHouse.log_file())
-}
-
-fn cleanup_legacy_clickhouse_pipe_log(paths: &RuntimePaths) {
-    let legacy_log = legacy_clickhouse_pipe_log_path(paths);
-    let should_remove = fs::metadata(&legacy_log)
-        .map(|metadata| metadata.is_file())
-        .unwrap_or(false);
-    if should_remove {
-        let _ = fs::remove_file(legacy_log);
-    }
-}
-
-fn log_path(paths: &RuntimePaths, service: Service) -> PathBuf {
-    match service {
-        Service::ClickHouse => clickhouse_internal_log_path(paths),
-        Service::Ingest | Service::Monitor | Service::Mcp => {
-            paths.logs_dir.join(service.log_file())
-        }
-    }
-}
-
-fn read_pid(path: &Path) -> Option<u32> {
-    let text = fs::read_to_string(path).ok()?;
-    text.trim().parse::<u32>().ok()
-}
-
-fn write_pid(path: &Path, pid: u32) -> Result<()> {
-    fs::write(path, format!("{}\n", pid))
-        .with_context(|| format!("failed to write pid file {}", path.display()))
-}
-
-fn is_pid_running(pid: u32) -> bool {
-    Command::new("kill")
-        .arg("-0")
-        .arg(pid.to_string())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .status()
-        .map(|status| status.success())
-        .unwrap_or(false)
-}
-
-fn ensure_pid_fresh(path: &Path) {
-    if let Some(pid) = read_pid(path) {
-        if !is_pid_running(pid) {
-            let _ = fs::remove_file(path);
-        }
-    }
-}
-
-fn service_running(paths: &RuntimePaths, service: Service) -> Option<u32> {
-    let path = pid_path(paths, service);
-    ensure_pid_fresh(&path);
-    let pid = read_pid(&path)?;
-    if is_pid_running(pid) {
-        Some(pid)
-    } else {
-        None
-    }
-}
-
-/// True when something is currently accepting connections on the central MCP
-/// socket. Used alongside (not instead of) the PID file: the daemon can
-/// outlive a deleted or stale PID file, and acting on the file alone either
-/// spawns a useless second daemon (`up`) or unlinks a live server's socket,
-/// orphaning it unreachably (`down`).
-fn central_socket_live(socket_path: &str) -> bool {
-    #[cfg(unix)]
-    {
-        std::os::unix::net::UnixStream::connect(socket_path).is_ok()
-    }
-    #[cfg(not(unix))]
-    {
-        let _ = socket_path;
-        false
-    }
-}
-
-fn stop_service(paths: &RuntimePaths, service: Service) -> Result<bool> {
-    let path = pid_path(paths, service);
-    let Some(pid) = read_pid(&path) else {
-        return Ok(false);
-    };
-
-    if !is_pid_running(pid) {
-        let _ = fs::remove_file(path);
-        return Ok(false);
-    }
-
-    let _ = Command::new("kill")
-        .arg(pid.to_string())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .status();
-
-    for _ in 0..20 {
-        if !is_pid_running(pid) {
-            let _ = fs::remove_file(&path);
-            return Ok(true);
-        }
-        std::thread::sleep(Duration::from_millis(200));
-    }
-
-    let _ = Command::new("kill")
-        .arg("-9")
-        .arg(pid.to_string())
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .status();
-    let _ = fs::remove_file(path);
-    Ok(true)
-}
-
-fn parse_config_flag(args: &[String]) -> Result<(Option<PathBuf>, Vec<String>)> {
-    let mut raw_config = None;
-    let mut rest = Vec::new();
-
-    let mut i = 0usize;
-    while i < args.len() {
-        if args[i] == "--config" {
-            if i + 1 >= args.len() {
-                bail!("--config requires a path");
-            }
-            raw_config = Some(PathBuf::from(args[i + 1].clone()));
-            i += 2;
-            continue;
-        }
-
-        rest.push(args[i].clone());
-        i += 1;
-    }
-
-    Ok((raw_config, rest))
-}
-
-fn load_cfg(raw_config: Option<PathBuf>) -> Result<(PathBuf, AppConfig)> {
-    let config_path = moraine_config::resolve_config_path(raw_config);
-    let cfg = moraine_config::load_config(&config_path)
-        .with_context(|| format!("failed to load config {}", config_path.display()))?;
-    Ok((config_path, cfg))
-}
-
-fn clickhouse_ports_from_url(cfg: &AppConfig) -> Result<(u16, u16, u16)> {
-    let parsed = Url::parse(&cfg.clickhouse.url)
-        .with_context(|| format!("invalid clickhouse.url '{}'", cfg.clickhouse.url))?;
-    let http_port = parsed.port_or_known_default().ok_or_else(|| {
-        anyhow!(
-            "clickhouse.url '{}' must include a known port",
-            cfg.clickhouse.url
-        )
-    })?;
-    let tcp_port = http_port
-        .checked_add(877)
-        .ok_or_else(|| anyhow!("derived clickhouse tcp port overflow from {}", http_port))?;
-    let interserver_http_port = http_port.checked_add(886).ok_or_else(|| {
-        anyhow!(
-            "derived clickhouse interserver port overflow from {}",
-            http_port
-        )
-    })?;
-    Ok((http_port, tcp_port, interserver_http_port))
-}
-
-fn clickhouse_url_is_local(cfg: &AppConfig) -> Result<bool> {
-    let parsed = Url::parse(&cfg.clickhouse.url)
-        .with_context(|| format!("invalid clickhouse.url '{}'", cfg.clickhouse.url))?;
-    let Some(host) = parsed.host_str() else {
-        bail!(
-            "clickhouse.url '{}' must include a host",
-            cfg.clickhouse.url
-        );
-    };
-
-    Ok(matches!(
-        host.to_ascii_lowercase().as_str(),
-        "localhost" | "127.0.0.1" | "::1" | "[::1]" | "0.0.0.0" | "::" | "[::]"
-    ))
-}
-
-fn materialize_clickhouse_config(cfg: &AppConfig, paths: &RuntimePaths) -> Result<()> {
-    let (http_port, tcp_port, interserver_http_port) = clickhouse_ports_from_url(cfg)?;
-    let rendered_clickhouse = CLICKHOUSE_TEMPLATE
-        .replace("__MORAINE_HOME__", &cfg.runtime.root_dir)
-        .replace("__CLICKHOUSE_HTTP_PORT__", &http_port.to_string())
-        .replace("__CLICKHOUSE_TCP_PORT__", &tcp_port.to_string())
-        .replace(
-            "__CLICKHOUSE_INTERSERVER_HTTP_PORT__",
-            &interserver_http_port.to_string(),
-        );
-    let rendered_users = USERS_TEMPLATE.replace("__MORAINE_HOME__", &cfg.runtime.root_dir);
-
-    fs::write(&paths.clickhouse_config, rendered_clickhouse).with_context(|| {
-        format!(
-            "failed writing clickhouse config {}",
-            paths.clickhouse_config.display()
-        )
-    })?;
-    fs::write(&paths.clickhouse_users, rendered_users).with_context(|| {
-        format!(
-            "failed writing users config {}",
-            paths.clickhouse_users.display()
-        )
-    })?;
-
-    Ok(())
-}
-
-fn managed_clickhouse_bin(paths: &RuntimePaths, binary: &str) -> PathBuf {
-    paths.managed_clickhouse_dir.join("bin").join(binary)
-}
-
-fn managed_clickhouse_checksum_file(paths: &RuntimePaths) -> PathBuf {
-    paths.managed_clickhouse_dir.join("SHA256")
-}
-
-fn detect_host_target() -> Result<&'static str> {
-    match (std::env::consts::OS, std::env::consts::ARCH) {
-        ("macos", "x86_64") => Ok("x86_64-apple-darwin"),
-        ("macos", "aarch64") => Ok("aarch64-apple-darwin"),
-        ("linux", "x86_64") => Ok("x86_64-unknown-linux-gnu"),
-        ("linux", "aarch64") => Ok("aarch64-unknown-linux-gnu"),
-        (os, arch) => bail!(
-            "unsupported platform for managed ClickHouse: {} {}",
-            os,
-            arch
-        ),
-    }
-}
-
-fn clickhouse_asset_for_target(version: &str, target: &str) -> Result<ClickHouseAsset> {
-    if version != DEFAULT_CLICKHOUSE_TAG {
-        bail!(
-            "unsupported ClickHouse version {}; this build supports {}",
-            version,
-            DEFAULT_CLICKHOUSE_TAG
-        );
-    }
-
-    match target {
-        "x86_64-apple-darwin" => Ok(ClickHouseAsset {
-            url: CH_URL_MACOS_X86_64,
-            sha256: CH_SHA_MACOS_X86_64,
-            is_archive: false,
-        }),
-        "aarch64-apple-darwin" => Ok(ClickHouseAsset {
-            url: CH_URL_MACOS_AARCH64,
-            sha256: CH_SHA_MACOS_AARCH64,
-            is_archive: false,
-        }),
-        "x86_64-unknown-linux-gnu" => Ok(ClickHouseAsset {
-            url: CH_URL_LINUX_X86_64,
-            sha256: CH_SHA_LINUX_X86_64,
-            is_archive: true,
-        }),
-        "aarch64-unknown-linux-gnu" => Ok(ClickHouseAsset {
-            url: CH_URL_LINUX_AARCH64,
-            sha256: CH_SHA_LINUX_AARCH64,
-            is_archive: true,
-        }),
-        other => bail!("unsupported ClickHouse target: {}", other),
-    }
-}
-
-fn clickhouse_asset_for_host(version: &str) -> Result<ClickHouseAsset> {
-    clickhouse_asset_for_target(version, detect_host_target()?)
-}
-
-async fn download_to_path(url: &str, dest: &Path, label: &str) -> Result<()> {
-    let client = Client::new();
-    let mut response = client
-        .get(url)
-        .send()
-        .await
-        .with_context(|| format!("failed to download {}", url))?
-        .error_for_status()
-        .with_context(|| format!("download failed for {}", url))?;
-
-    let total = response.content_length();
-    let show_progress = std::io::stderr().is_terminal();
-
-    let mut file = std::fs::File::create(dest)
-        .with_context(|| format!("failed writing {}", dest.display()))?;
-    let mut downloaded: u64 = 0;
-    let mut last_render = Instant::now();
-
-    while let Some(chunk) = response
-        .chunk()
-        .await
-        .with_context(|| format!("failed reading response body for {}", url))?
-    {
-        file.write_all(&chunk)
-            .with_context(|| format!("failed writing {}", dest.display()))?;
-        downloaded += chunk.len() as u64;
-
-        if show_progress && last_render.elapsed() >= Duration::from_millis(150) {
-            render_download_progress(label, downloaded, total, false);
-            last_render = Instant::now();
-        }
-    }
-
-    if show_progress {
-        render_download_progress(label, downloaded, total, true);
-    }
-
-    Ok(())
-}
-
-fn render_download_progress(label: &str, done: u64, total: Option<u64>, done_flag: bool) {
-    const MIB: f64 = 1024.0 * 1024.0;
-    let done_mb = done as f64 / MIB;
-    match total {
-        Some(t) if t > 0 => {
-            let total_mb = t as f64 / MIB;
-            let pct = ((done as f64 / t as f64) * 100.0).min(100.0);
-            eprint!("\r  {label}: {done_mb:>6.1} / {total_mb:>6.1} MiB  ({pct:>5.1}%)");
-        }
-        _ => {
-            eprint!("\r  {label}: {done_mb:>6.1} MiB");
-        }
-    }
-    if done_flag {
-        eprintln!();
-    } else {
-        std::io::stderr().flush().ok();
-    }
-}
-
-fn sha256_hex(path: &Path) -> Result<String> {
-    let mut file = std::fs::File::open(path)
-        .with_context(|| format!("failed opening {} for checksum", path.display()))?;
-    let mut hasher = Sha256::new();
-    let mut buf = [0u8; 64 * 1024];
-
-    loop {
-        let n = std::io::Read::read(&mut file, &mut buf)?;
-        if n == 0 {
-            break;
-        }
-        hasher.update(&buf[..n]);
-    }
-
-    let digest = hasher.finalize();
-    Ok(digest.iter().map(|b| format!("{:02x}", b)).collect())
-}
-
-fn path_ends_with_components(path: &Path, suffix: &[&str]) -> bool {
-    let mut components = path.components().rev();
-    for expected in suffix.iter().rev() {
-        match components
-            .next()
-            .and_then(|component| component.as_os_str().to_str())
-        {
-            Some(component) if component == *expected => {}
-            _ => return false,
-        }
-    }
-
-    true
-}
-
-fn find_file_ending_with(root: &Path, suffix: &[&str]) -> Result<Option<PathBuf>> {
-    let mut stack = vec![root.to_path_buf()];
-
-    while let Some(dir) = stack.pop() {
-        let entries = match fs::read_dir(&dir) {
-            Ok(entries) => entries,
-            Err(_) => continue,
-        };
-
-        for entry in entries {
-            let entry = match entry {
-                Ok(entry) => entry,
-                Err(_) => continue,
-            };
-
-            let path = entry.path();
-            if path.is_dir() {
-                stack.push(path);
-                continue;
-            }
-
-            if path_ends_with_components(&path, suffix) {
-                return Ok(Some(path));
-            }
-        }
-    }
-
-    Ok(None)
-}
-
-fn find_file_named(root: &Path, name: &str) -> Result<Option<PathBuf>> {
-    let mut stack = vec![root.to_path_buf()];
-
-    while let Some(dir) = stack.pop() {
-        let entries = match fs::read_dir(&dir) {
-            Ok(entries) => entries,
-            Err(_) => continue,
-        };
-
-        for entry in entries {
-            let entry = match entry {
-                Ok(entry) => entry,
-                Err(_) => continue,
-            };
-
-            let path = entry.path();
-            if path.is_dir() {
-                stack.push(path);
-                continue;
-            }
-
-            if path
-                .file_name()
-                .and_then(|s| s.to_str())
-                .is_some_and(|s| s == name)
-            {
-                return Ok(Some(path));
-            }
-        }
-    }
-
-    Ok(None)
-}
-
-#[cfg(unix)]
-fn make_executable(path: &Path) -> Result<()> {
-    let mut perms = fs::metadata(path)?.permissions();
-    perms.set_mode(0o755);
-    fs::set_permissions(path, perms)?;
-    Ok(())
-}
-
-#[cfg(not(unix))]
-fn make_executable(_path: &Path) -> Result<()> {
-    Ok(())
-}
-
-#[cfg(unix)]
-fn ensure_symlink(target: &Path, link: &Path) -> Result<()> {
-    if link.exists() {
-        let _ = fs::remove_file(link);
-    }
-    symlink(target, link).with_context(|| {
-        format!(
-            "failed creating symlink {} -> {}",
-            link.display(),
-            target.display()
-        )
-    })
-}
-
-#[cfg(not(unix))]
-fn ensure_symlink(target: &Path, link: &Path) -> Result<()> {
-    fs::copy(target, link)?;
-    Ok(())
-}
-
-async fn install_managed_clickhouse(
-    paths: &RuntimePaths,
-    version: &str,
-    force: bool,
-) -> Result<PathBuf> {
-    let asset = clickhouse_asset_for_host(version)?;
-
-    let bin_dir = paths.managed_clickhouse_dir.join("bin");
-    let clickhouse = bin_dir.join("clickhouse");
-    let clickhouse_server = bin_dir.join("clickhouse-server");
-
-    if clickhouse_server.exists() && !force {
-        return Ok(clickhouse_server);
-    }
-
-    fs::create_dir_all(paths.root.join("tmp"))
-        .with_context(|| format!("failed to create {}", paths.root.join("tmp").display()))?;
-
-    let stamp = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .unwrap_or_default()
-        .as_millis();
-
-    let download = paths
-        .root
-        .join("tmp")
-        .join(format!("clickhouse-download-{}", stamp));
-    let extract_dir = paths
-        .root
-        .join("tmp")
-        .join(format!("clickhouse-extract-{}", stamp));
-
-    download_to_path(asset.url, &download, &format!("ClickHouse {version}")).await?;
-
-    let digest = sha256_hex(&download)?;
-    if digest != asset.sha256 {
-        bail!(
-            "managed ClickHouse checksum mismatch: expected {}, got {}",
-            asset.sha256,
-            digest
-        );
-    }
-
-    fs::create_dir_all(&extract_dir)?;
-
-    let staged_binary = if asset.is_archive {
-        let status = Command::new("tar")
-            .env("LC_ALL", "C")
-            .arg("-xzf")
-            .arg(&download)
-            .arg("-C")
-            .arg(&extract_dir)
-            .status()
-            .context("failed to run tar while installing ClickHouse")?;
-        if !status.success() {
-            bail!("failed to extract ClickHouse archive");
-        }
-
-        find_file_ending_with(&extract_dir, &["usr", "bin", "clickhouse"])?
-            .or(find_file_named(&extract_dir, "clickhouse")?)
-            .ok_or_else(|| anyhow!("extracted ClickHouse archive missing clickhouse binary"))?
-    } else {
-        download.clone()
-    };
-
-    fs::create_dir_all(&bin_dir)
-        .with_context(|| format!("failed creating {}", bin_dir.display()))?;
-
-    fs::copy(&staged_binary, &clickhouse)
-        .with_context(|| format!("failed writing {}", clickhouse.display()))?;
-    make_executable(&clickhouse)?;
-
-    ensure_symlink(&clickhouse, &clickhouse_server)?;
-    ensure_symlink(&clickhouse, &bin_dir.join("clickhouse-client"))?;
-
-    fs::write(
-        paths.managed_clickhouse_dir.join("VERSION"),
-        format!("{}\n", version),
-    )?;
-    fs::write(
-        managed_clickhouse_checksum_file(paths),
-        format!("{}\n", digest),
-    )?;
-
-    let _ = fs::remove_file(download);
-    let _ = fs::remove_dir_all(extract_dir);
-
-    Ok(clickhouse_server)
-}
-
-fn clickhouse_from_path_available() -> bool {
-    Command::new("clickhouse-server")
-        .arg("--version")
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .status()
-        .map(|s| s.success())
-        .unwrap_or(false)
-}
-
-async fn resolve_clickhouse_server_command(
-    cfg: &AppConfig,
-    paths: &RuntimePaths,
-) -> Result<PathBuf> {
-    let managed = managed_clickhouse_bin(paths, "clickhouse-server");
-    if managed.exists() {
-        return Ok(managed);
-    }
-
-    if clickhouse_from_path_available() {
-        return Ok(PathBuf::from("clickhouse-server"));
-    }
-
-    let version = &cfg.runtime.clickhouse_version;
-    let should_install = if cfg.runtime.clickhouse_auto_install {
-        eprintln!(
-            "managed ClickHouse not found; auto-installing {version}.\n\
-             one-time ~175 MiB download + extract (progress shown below).\n\
-             set runtime.clickhouse_auto_install = false in your config to disable.",
-        );
-        true
-    } else {
-        prompt_install_clickhouse(version)?
-    };
-
-    if should_install {
-        install_managed_clickhouse(paths, version, false).await?;
-        let managed = managed_clickhouse_bin(paths, "clickhouse-server");
-        if managed.exists() {
-            return Ok(managed);
-        }
-    }
-
-    bail!(
-        "clickhouse-server is not installed or not on PATH (managed install dir: {})",
-        paths.managed_clickhouse_dir.display()
-    )
-}
-
-fn prompt_install_clickhouse(version: &str) -> Result<bool> {
-    let stdin = std::io::stdin();
-    if !stdin.is_terminal() || !std::io::stderr().is_terminal() {
-        bail!(
-            "managed ClickHouse is not installed and runtime.clickhouse_auto_install = false.\n\
-             cannot prompt in non-interactive mode.\n\
-             remediation:\n\
-             - run `moraine clickhouse install` explicitly, or\n\
-             - set runtime.clickhouse_auto_install = true in your config"
-        );
-    }
-
-    loop {
-        eprint!("managed ClickHouse {version} is not installed. install now? [Y/n] ");
-        std::io::stderr().flush().ok();
-
-        let mut input = String::new();
-        stdin
-            .read_line(&mut input)
-            .context("failed to read confirmation from stdin")?;
-
-        match input.trim().to_ascii_lowercase().as_str() {
-            "" | "y" | "yes" => return Ok(true),
-            "n" | "no" => return Ok(false),
-            _ => eprintln!("please answer 'y' or 'n'."),
-        }
-    }
-}
-
-async fn wait_for_clickhouse(cfg: &AppConfig) -> Result<()> {
-    let client = ClickHouseClient::new(cfg.clickhouse.clone())?;
-    let timeout = Duration::from_secs_f64(cfg.runtime.clickhouse_start_timeout_seconds.max(1.0));
-    let interval = Duration::from_millis(cfg.runtime.healthcheck_interval_ms.max(100));
-    let start = Instant::now();
-
-    loop {
-        if client.ping().await.is_ok() {
-            return Ok(());
-        }
-
-        if start.elapsed() >= timeout {
-            bail!(
-                "clickhouse did not become healthy within {:.1}s",
-                timeout.as_secs_f64()
-            );
-        }
-
-        sleep(interval).await;
-    }
-}
-
-async fn start_clickhouse(cfg: &AppConfig, paths: &RuntimePaths) -> Result<StartOutcome> {
-    if let Some(pid) = service_running(paths, Service::ClickHouse) {
-        return Ok(StartOutcome {
-            service: Service::ClickHouse,
-            state: StartState::AlreadyRunning,
-            pid: Some(pid),
-            log_path: Some(log_path(paths, Service::ClickHouse).display().to_string()),
-        });
-    }
-
-    let url_is_local = clickhouse_url_is_local(cfg)?;
-    let client = ClickHouseClient::new(cfg.clickhouse.clone())?;
-    match client.ping().await {
-        Ok(()) => {
-            return Ok(StartOutcome {
-                service: Service::ClickHouse,
-                state: StartState::AlreadyServing,
-                pid: None,
-                log_path: None,
-            });
-        }
-        Err(err) if !url_is_local => {
-            bail!(
-                "clickhouse.url '{}' points at a non-local endpoint, but the endpoint is not healthy: {err}. start or repair that ClickHouse endpoint before running `moraine up`; Moraine only starts managed ClickHouse for local URLs",
-                cfg.clickhouse.url
-            );
-        }
-        Err(_) => {}
-    }
-
-    cleanup_legacy_clickhouse_pipe_log(paths);
-
-    let server_bin = resolve_clickhouse_server_command(cfg, paths).await?;
-
-    materialize_clickhouse_config(cfg, paths)?;
-
-    let child = Command::new(&server_bin)
-        .arg("--config-file")
-        .arg(&paths.clickhouse_config)
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .spawn()
-        .with_context(|| format!("failed to start {}", server_bin.display()))?;
-
-    write_pid(&pid_path(paths, Service::ClickHouse), child.id())?;
-
-    wait_for_clickhouse(cfg).await?;
-    Ok(StartOutcome {
-        service: Service::ClickHouse,
-        state: StartState::Started,
-        pid: Some(child.id()),
-        log_path: Some(log_path(paths, Service::ClickHouse).display().to_string()),
-    })
-}
-
-async fn run_foreground_clickhouse(cfg: &AppConfig, paths: &RuntimePaths) -> Result<ExitCode> {
-    ensure_runtime_dirs(paths)?;
-    let server_bin = resolve_clickhouse_server_command(cfg, paths).await?;
-    materialize_clickhouse_config(cfg, paths)?;
-
-    let status = Command::new(server_bin)
-        .arg("--config-file")
-        .arg(&paths.clickhouse_config)
-        .status()
-        .context("failed to run clickhouse-server")?;
-
-    Ok(ExitCode::from(status.code().unwrap_or(1) as u8))
-}
-
-fn env_flag_enabled(key: &str) -> bool {
-    std::env::var(key)
-        .ok()
-        .map(|value| {
-            matches!(
-                value.trim().to_ascii_lowercase().as_str(),
-                "1" | "true" | "yes" | "on"
-            )
-        })
-        .unwrap_or(false)
-}
-
-fn source_tree_mode_enabled() -> bool {
-    env_flag_enabled("MORAINE_SOURCE_TREE_MODE")
-}
-
-#[derive(Debug, Clone)]
-struct ServiceBinaryProbe {
-    source: &'static str,
-    path: PathBuf,
-}
-
-#[derive(Debug, Clone)]
-struct ServiceBinaryResolution {
-    binary_name: String,
-    resolved_path: Option<PathBuf>,
-    checked_paths: Vec<ServiceBinaryProbe>,
-}
-
-fn resolve_service_binary(service: Service, paths: &RuntimePaths) -> ServiceBinaryResolution {
-    let name = service.binary_name().unwrap_or(service.name()).to_string();
-    let mut checked_paths = Vec::new();
-
-    let mut check = |source: &'static str, path: PathBuf| {
-        if path.exists() {
-            Some(path)
-        } else {
-            checked_paths.push(ServiceBinaryProbe { source, path });
-            None
-        }
-    };
-
-    if let Ok(dir) = std::env::var("MORAINE_SERVICE_BIN_DIR") {
-        if let Some(path) = check("MORAINE_SERVICE_BIN_DIR", PathBuf::from(dir).join(&name)) {
-            return ServiceBinaryResolution {
-                binary_name: name,
-                resolved_path: Some(path),
-                checked_paths,
-            };
-        }
-    }
-
-    if let Some(path) = check("runtime.service_bin_dir", paths.service_bin_dir.join(&name)) {
-        return ServiceBinaryResolution {
-            binary_name: name,
-            resolved_path: Some(path),
-            checked_paths,
-        };
-    }
-
-    if let Ok(exe) = std::env::current_exe() {
-        if let Some(dir) = exe.parent() {
-            if let Some(path) = check("moraine sibling", dir.join(&name)) {
-                return ServiceBinaryResolution {
-                    binary_name: name,
-                    resolved_path: Some(path),
-                    checked_paths,
-                };
-            }
-        }
-    }
-
-    if source_tree_mode_enabled() {
-        if let Some(project_bin) = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .and_then(|p| p.parent())
-            .map(|p| p.join("target").join("debug").join(&name))
-        {
-            if let Some(path) = check("source-tree mode target/debug", project_bin) {
-                return ServiceBinaryResolution {
-                    binary_name: name,
-                    resolved_path: Some(path),
-                    checked_paths,
-                };
-            }
-        }
-    }
-
-    ServiceBinaryResolution {
-        binary_name: name,
-        resolved_path: None,
-        checked_paths,
-    }
-}
-
-fn require_service_binary(service: Service, paths: &RuntimePaths) -> Result<PathBuf> {
-    let resolution = resolve_service_binary(service, paths);
-    if let Some(path) = resolution.resolved_path {
-        return Ok(path);
-    }
-
-    let checked = if resolution.checked_paths.is_empty() {
-        "- (no probe paths)".to_string()
-    } else {
-        resolution
-            .checked_paths
-            .iter()
-            .map(|probe| format!("- {} ({})", probe.path.display(), probe.source))
-            .collect::<Vec<_>>()
-            .join("\n")
-    };
-
-    bail!(
-        "required service binary `{}` for `{}` was not found.\nchecked:\n{}\nremediation:\n- install Moraine service binaries so `{}` exists under `runtime.service_bin_dir` (`{}`)\n- or set `MORAINE_SERVICE_BIN_DIR` to a directory containing `{}`\n- for source builds run `cargo build --workspace --locked` and set `MORAINE_SOURCE_TREE_MODE=1`\n`moraine` does not fall back to PATH for service binaries.",
-        resolution.binary_name,
-        service.name(),
-        checked,
-        resolution.binary_name,
-        paths.service_bin_dir.display(),
-        resolution.binary_name
-    );
-}
-
-fn preflight_required_service_binaries(services: &[Service], paths: &RuntimePaths) -> Result<()> {
-    for service in services {
-        if *service == Service::ClickHouse {
-            continue;
-        }
-        require_service_binary(*service, paths)?;
-    }
-    Ok(())
-}
-
-fn contains_flag(args: &[String], flag: &str) -> bool {
-    args.iter().any(|arg| arg == flag)
-}
-
-fn monitor_dist_candidate(root: &Path) -> PathBuf {
-    root.join("web").join("monitor").join("dist")
-}
-
-const MONITOR_DIST_ENV_KEYS: &[&str] = &["MORAINE_MONITOR_DIST", "MORAINE_MONITOR_STATIC_DIR"];
-
-fn resolve_monitor_static_dir(paths: &RuntimePaths) -> Option<PathBuf> {
-    for key in MONITOR_DIST_ENV_KEYS {
-        if let Ok(value) = std::env::var(key) {
-            let trimmed = value.trim();
-            if trimmed.is_empty() {
-                continue;
-            }
-            let path = PathBuf::from(trimmed);
-            if path.exists() {
-                return Some(path);
-            }
-        }
-    }
-
-    if let Ok(exe) = std::env::current_exe() {
-        if let Some(bin_dir) = exe.parent() {
-            if let Some(bundle_root) = bin_dir.parent() {
-                let candidate = monitor_dist_candidate(bundle_root);
-                if candidate.exists() {
-                    return Some(candidate);
-                }
-            }
-        }
-    }
-
-    if let Some(bundle_root) = paths.service_bin_dir.parent() {
-        let candidate = monitor_dist_candidate(bundle_root);
-        if candidate.exists() {
-            return Some(candidate);
-        }
-    }
-
-    if source_tree_mode_enabled() {
-        if let Some(dev_path) = Path::new(env!("CARGO_MANIFEST_DIR"))
-            .parent()
-            .and_then(|p| p.parent())
-            .map(PathBuf::from)
-        {
-            let candidate = monitor_dist_candidate(&dev_path);
-            if candidate.exists() {
-                return Some(candidate);
-            }
-        }
-    }
-
-    None
-}
-
-fn service_args_with_defaults(
-    service: Service,
-    cfg_path: &Path,
-    cfg: &AppConfig,
-    paths: &RuntimePaths,
-    passthrough: &[String],
-) -> Vec<String> {
-    let mut args = Vec::new();
-
-    if !contains_flag(passthrough, "--config") {
-        args.push("--config".to_string());
-        args.push(cfg_path.to_string_lossy().to_string());
-    }
-
-    if service == Service::Monitor {
-        if !contains_flag(passthrough, "--host") {
-            args.push("--host".to_string());
-            args.push(cfg.monitor.host.clone());
-        }
-        if !contains_flag(passthrough, "--port") {
-            args.push("--port".to_string());
-            args.push(cfg.monitor.port.to_string());
-        }
-        if !contains_flag(passthrough, "--static-dir") {
-            if let Some(static_dir) = resolve_monitor_static_dir(paths) {
-                args.push("--static-dir".to_string());
-                args.push(static_dir.to_string_lossy().to_string());
-            }
-        }
-    }
-
-    args.extend(passthrough.iter().cloned());
-    args
-}
-
-fn start_background_service(
-    service: Service,
-    cfg_path: &Path,
-    cfg: &AppConfig,
-    paths: &RuntimePaths,
-    extra_args: &[String],
-) -> Result<StartOutcome> {
-    if service == Service::ClickHouse {
-        bail!("clickhouse is not managed by service launcher; use `moraine up`");
-    }
-
-    if let Some(pid) = service_running(paths, service) {
-        return Ok(StartOutcome {
-            service,
-            state: StartState::AlreadyRunning,
-            pid: Some(pid),
-            log_path: Some(log_path(paths, service).display().to_string()),
-        });
-    }
-
-    // The central MCP server can outlive its PID file (e.g. the file was
-    // deleted while the daemon survived). Probe the socket before spawning:
-    // a second daemon would just no-op-exit, and the write_pid below would
-    // then record that dead PID, breaking `status` and `down` for the live
-    // server.
-    if service == Service::Mcp && central_socket_live(&cfg.mcp.central_socket_path) {
-        return Ok(StartOutcome {
-            service,
-            state: StartState::AlreadyServing,
-            pid: None,
-            log_path: Some(log_path(paths, service).display().to_string()),
-        });
-    }
-
-    let binary = require_service_binary(service, paths)?;
-
-    let logfile = OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(log_path(paths, service))
-        .with_context(|| format!("failed to open {} log", service.name()))?;
-    let logfile_err = logfile
-        .try_clone()
-        .with_context(|| format!("failed to clone {} log", service.name()))?;
-
-    let args = service_args_with_defaults(service, cfg_path, cfg, paths, extra_args);
-
-    let child = Command::new(&binary)
-        .args(args)
-        // Background services never read stdin; closing it avoids inheriting the
-        // launcher's terminal/pipe fd (and a SIGHUP/blocking-read footgun for the
-        // central MCP server, which would otherwise hold an unused stdin handle).
-        .stdin(Stdio::null())
-        .stdout(Stdio::from(logfile))
-        .stderr(Stdio::from(logfile_err))
-        .spawn()
-        .with_context(|| format!("failed to start {}", service.name()))?;
-
-    write_pid(&pid_path(paths, service), child.id())?;
-    Ok(StartOutcome {
-        service,
-        state: StartState::Started,
-        pid: Some(child.id()),
-        log_path: Some(log_path(paths, service).display().to_string()),
-    })
-}
-
-async fn run_foreground_service(
-    service: Service,
-    cfg_path: &Path,
-    cfg: &AppConfig,
-    paths: &RuntimePaths,
-    passthrough_args: &[String],
-) -> Result<ExitCode> {
-    if service == Service::ClickHouse {
-        return run_foreground_clickhouse(cfg, paths).await;
-    }
-
-    let binary = require_service_binary(service, paths)?;
-    let args = service_args_with_defaults(service, cfg_path, cfg, paths, passthrough_args);
-
-    let status = Command::new(binary)
-        .args(args)
-        .status()
-        .with_context(|| format!("failed to run {}", service.name()))?;
-
-    Ok(ExitCode::from(status.code().unwrap_or(1) as u8))
-}
-
-async fn cmd_db_migrate(cfg: &AppConfig) -> Result<MigrationOutcome> {
-    let ch = ClickHouseClient::new(cfg.clickhouse.clone())?;
-    let applied = ch.run_migrations().await?;
-    Ok(MigrationOutcome { applied })
-}
-
-async fn cmd_db_doctor(cfg: &AppConfig) -> Result<DoctorReport> {
-    let ch = ClickHouseClient::new(cfg.clickhouse.clone())?;
-    ch.doctor_report().await
-}
-
-async fn query_heartbeat(cfg: &AppConfig) -> Result<Option<HeartbeatRow>> {
-    let ch = ClickHouseClient::new(cfg.clickhouse.clone())?;
-    let db = quote_identifier(&cfg.clickhouse.database);
-    let query = format!(
-        "SELECT \
-            toString(max(ts)) AS latest, \
-            toUInt64(argMax(queue_depth, ts)) AS queue_depth, \
-            toUInt64(argMax(files_active, ts)) AS files_active, \
-            toString(argMax(watcher_backend, ts)) AS watcher_backend, \
-            toUInt64(argMax(watcher_error_count, ts)) AS watcher_error_count, \
-            toUInt64(argMax(watcher_reset_count, ts)) AS watcher_reset_count, \
-            toUInt64(argMax(watcher_last_reset_unix_ms, ts)) AS watcher_last_reset_unix_ms \
-         FROM {db}.ingest_heartbeats"
-    );
-
-    match ch.query_json_data::<HeartbeatRow>(&query, None).await {
-        Ok(rows) => Ok(rows.into_iter().next()),
-        Err(_) => {
-            let legacy_query = format!(
-                "SELECT toString(max(ts)) AS latest, toUInt64(argMax(queue_depth, ts)) AS queue_depth, toUInt64(argMax(files_active, ts)) AS files_active FROM {db}.ingest_heartbeats"
-            );
-            let rows: Vec<LegacyHeartbeatRow> = ch.query_json_data(&legacy_query, None).await?;
-            Ok(rows.into_iter().next().map(|row| HeartbeatRow {
-                latest: row.latest,
-                queue_depth: row.queue_depth,
-                files_active: row.files_active,
-                watcher_backend: "unknown".to_string(),
-                watcher_error_count: 0,
-                watcher_reset_count: 0,
-                watcher_last_reset_unix_ms: 0,
-            }))
-        }
-    }
-}
-
-fn quote_identifier(value: &str) -> String {
-    format!("`{}`", value.replace('`', "``"))
-}
-
-fn managed_clickhouse_version(paths: &RuntimePaths) -> Option<String> {
-    fs::read_to_string(paths.managed_clickhouse_dir.join("VERSION"))
-        .ok()
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-}
-
-fn managed_clickhouse_checksum(paths: &RuntimePaths) -> Option<String> {
-    fs::read_to_string(managed_clickhouse_checksum_file(paths))
-        .ok()
-        .map(|s| s.trim().to_string())
-        .filter(|s| !s.is_empty())
-}
-
-fn managed_clickhouse_checksum_state(cfg: &AppConfig, paths: &RuntimePaths) -> String {
-    let Some(stored) = managed_clickhouse_checksum(paths) else {
-        return "unknown (missing checksum metadata)".to_string();
-    };
-
-    let expected = match clickhouse_asset_for_host(&cfg.runtime.clickhouse_version) {
-        Ok(asset) => asset.sha256,
-        Err(exc) => return format!("unknown ({})", exc),
-    };
-
-    if stored == expected {
-        "verified".to_string()
-    } else {
-        format!("mismatch (expected {}, got {})", expected, stored)
-    }
-}
-
-fn active_clickhouse_source(paths: &RuntimePaths) -> (&'static str, Option<PathBuf>) {
-    let managed = managed_clickhouse_bin(paths, "clickhouse-server");
-    if managed.exists() {
-        return ("managed", Some(managed));
-    }
-    if clickhouse_from_path_available() {
-        return ("path", Some(PathBuf::from("clickhouse-server")));
-    }
-    ("missing", None)
-}
-
-fn service_runtime_running(services: &[ServiceRuntimeStatus], service: Service) -> bool {
-    services
-        .iter()
-        .find(|row| row.service == service)
-        .and_then(|row| row.pid)
-        .is_some()
-}
-
-fn clickhouse_runtime_running(services: &[ServiceRuntimeStatus]) -> bool {
-    service_runtime_running(services, Service::ClickHouse)
-}
-
-fn monitor_runtime_running(services: &[ServiceRuntimeStatus]) -> bool {
-    service_runtime_running(services, Service::Monitor)
-}
-
-fn format_http_url(host: &str, port: u16) -> String {
-    if host.contains(':') && !(host.starts_with('[') && host.ends_with(']')) {
-        format!("http://[{host}]:{port}")
-    } else {
-        format!("http://{host}:{port}")
-    }
-}
-
-fn monitor_runtime_url(cfg: &AppConfig) -> String {
-    format_http_url(&cfg.monitor.host, cfg.monitor.port)
-}
-
-fn build_status_notes(
-    services: &[ServiceRuntimeStatus],
-    report: &DoctorReport,
-    clickhouse_url: &str,
-) -> Vec<String> {
-    let clickhouse_running = clickhouse_runtime_running(services);
-    let mut notes = Vec::new();
-
-    if report.clickhouse_healthy && !clickhouse_running {
-        notes.push(format!(
-            "database health checks query clickhouse.url ({clickhouse_url}); endpoint is healthy while managed clickhouse runtime is stopped"
-        ));
-    }
-
-    if !report.clickhouse_healthy && clickhouse_running {
-        notes.push(format!(
-            "managed clickhouse runtime is running, but health checks against clickhouse.url ({clickhouse_url}) are failing"
-        ));
-    }
-
-    notes
-}
-
-async fn cmd_status(paths: &RuntimePaths, cfg: &AppConfig) -> Result<StatusSnapshot> {
-    let services = [
-        Service::ClickHouse,
-        Service::Ingest,
-        Service::Monitor,
-        Service::Mcp,
-    ]
-    .iter()
-    .copied()
-    .map(|service| ServiceRuntimeStatus {
-        service,
-        pid: service_running(paths, service),
-    })
-    .collect::<Vec<_>>();
-    let managed_server = managed_clickhouse_bin(paths, "clickhouse-server");
-    let (source, source_path) = active_clickhouse_source(paths);
-    let report = cmd_db_doctor(cfg).await?;
-    let clickhouse_health_url = cfg.clickhouse.url.clone();
-    let status_notes = build_status_notes(&services, &report, &clickhouse_health_url);
-    let monitor_url = monitor_runtime_running(&services).then(|| monitor_runtime_url(cfg));
-    let heartbeat = match query_heartbeat(cfg).await {
-        Ok(Some(row)) => HeartbeatSnapshot::Available {
-            latest: row.latest,
-            queue_depth: row.queue_depth,
-            files_active: row.files_active,
-            watcher_backend: row.watcher_backend,
-            watcher_error_count: row.watcher_error_count,
-            watcher_reset_count: row.watcher_reset_count,
-            watcher_last_reset_unix_ms: row.watcher_last_reset_unix_ms,
-        },
-        Ok(None) => HeartbeatSnapshot::Unavailable,
-        Err(err) => HeartbeatSnapshot::Error {
-            message: err.to_string(),
-        },
-    };
-
-    Ok(StatusSnapshot {
-        services,
-        monitor_url,
-        managed_clickhouse_installed: managed_server.exists(),
-        managed_clickhouse_path: managed_server.display().to_string(),
-        managed_clickhouse_version: managed_clickhouse_version(paths),
-        clickhouse_active_source: source.to_string(),
-        clickhouse_active_source_path: source_path.map(|path| path.display().to_string()),
-        managed_clickhouse_checksum: managed_clickhouse_checksum_state(cfg, paths),
-        clickhouse_health_url,
-        status_notes,
-        doctor: report,
-        heartbeat,
-    })
-}
-
-fn tail_lines(path: &Path, lines: usize) -> Result<Vec<String>> {
-    const TAIL_READ_CHUNK_BYTES: usize = 8 * 1024;
-
-    if lines == 0 {
-        return Ok(Vec::new());
-    }
-
-    let mut file = fs::File::open(path)
-        .with_context(|| format!("failed to read log file {}", path.display()))?;
-    let mut position = file
-        .metadata()
-        .with_context(|| format!("failed to read log file {}", path.display()))?
-        .len();
-
-    let mut chunks: Vec<Vec<u8>> = Vec::new();
-    let mut scratch = vec![0_u8; TAIL_READ_CHUNK_BYTES];
-    let mut newline_count = 0usize;
-    while position > 0 {
-        let read_len = (position as usize).min(TAIL_READ_CHUNK_BYTES);
-        position -= read_len as u64;
-        file.seek(SeekFrom::Start(position))
-            .with_context(|| format!("failed to read log file {}", path.display()))?;
-        file.read_exact(&mut scratch[..read_len])
-            .with_context(|| format!("failed to read log file {}", path.display()))?;
-        newline_count += scratch[..read_len]
-            .iter()
-            .filter(|byte| **byte == b'\n')
-            .count();
-        chunks.push(scratch[..read_len].to_vec());
-        if newline_count > lines {
-            break;
-        }
-    }
-
-    let total_len = chunks.iter().map(Vec::len).sum();
-    let mut bytes = Vec::with_capacity(total_len);
-    for chunk in chunks.iter().rev() {
-        bytes.extend_from_slice(chunk);
-    }
-
-    let start = if position > 0 {
-        bytes
-            .iter()
-            .position(|byte| *byte == b'\n')
-            .map_or(bytes.len(), |idx| idx + 1)
-    } else {
-        0
-    };
-    let content = std::str::from_utf8(&bytes[start..])
-        .with_context(|| format!("failed to decode log file {} as utf-8", path.display()))?;
-    let mut collected = content
-        .lines()
-        .rev()
-        .take(lines)
-        .map(ToString::to_string)
-        .collect::<Vec<_>>();
-    collected.reverse();
-    Ok(collected)
-}
-
-fn collect_logs(
-    paths: &RuntimePaths,
-    service: Option<Service>,
-    lines: usize,
-) -> Result<LogsSnapshot> {
-    let targets = match service {
-        Some(svc) => vec![svc],
-        None => vec![
-            Service::ClickHouse,
-            Service::Ingest,
-            Service::Monitor,
-            Service::Mcp,
-        ],
-    };
-
-    let mut sections = Vec::new();
-    for svc in targets {
-        let path = log_path(paths, svc);
-        let path_string = path.display().to_string();
-        if !path.exists() {
-            sections.push(ServiceLogSection {
-                service: svc,
-                path: path_string,
-                exists: false,
-                lines: Vec::new(),
-            });
-            continue;
-        }
-        sections.push(ServiceLogSection {
-            service: svc,
-            path: path_string,
-            exists: true,
-            lines: tail_lines(&path, lines)?,
-        });
-    }
-
-    Ok(LogsSnapshot {
-        requested_lines: lines,
-        sections,
-    })
-}
-
-fn selected_up_services(args: &UpArgs, cfg: &AppConfig) -> Vec<Service> {
-    let mut services = Vec::new();
-    if !args.no_ingest {
-        services.push(Service::Ingest);
-    }
-    if args.monitor || cfg.runtime.start_monitor_on_up {
-        services.push(Service::Monitor);
-    }
-    // The up-managed Mcp service is always the shared central MCP server
-    // (`--serve socket`); the legacy stdio daemon variant is gone. `--mcp` and
-    // the deprecated runtime.start_mcp_on_up alias force it on even when
-    // mcp.start_central_on_up is disabled.
-    if args.mcp || cfg.runtime.start_mcp_on_up || cfg.mcp.start_central_on_up {
-        services.push(Service::Mcp);
-    }
-    services
-}
-
-async fn cmd_clickhouse_install(
-    paths: &RuntimePaths,
-    version: &str,
-    force: bool,
-) -> Result<PathBuf> {
-    ensure_runtime_dirs(paths)?;
-    let installed = install_managed_clickhouse(paths, version, force).await?;
-    Ok(installed)
-}
-
-fn cmd_clickhouse_status(cfg: &AppConfig, paths: &RuntimePaths) -> ClickhouseStatusSnapshot {
-    let clickhouse = managed_clickhouse_bin(paths, "clickhouse");
-    let clickhouse_server = managed_clickhouse_bin(paths, "clickhouse-server");
-    let clickhouse_client = managed_clickhouse_bin(paths, "clickhouse-client");
-    let (active_source, active_source_path) = active_clickhouse_source(paths);
-
-    ClickhouseStatusSnapshot {
-        managed_root: paths.managed_clickhouse_dir.display().to_string(),
-        clickhouse_exists: clickhouse.exists(),
-        clickhouse_server_exists: clickhouse_server.exists(),
-        clickhouse_client_exists: clickhouse_client.exists(),
-        expected_version: cfg.runtime.clickhouse_version.clone(),
-        active_source: active_source.to_string(),
-        active_source_path: active_source_path.map(|path| path.display().to_string()),
-        checksum_state: managed_clickhouse_checksum_state(cfg, paths),
-        installed_version: managed_clickhouse_version(paths),
-    }
-}
-
-fn cmd_clickhouse_uninstall(paths: &RuntimePaths) -> Result<String> {
-    if paths.managed_clickhouse_dir.exists() {
-        fs::remove_dir_all(&paths.managed_clickhouse_dir).with_context(|| {
-            format!("failed removing {}", paths.managed_clickhouse_dir.display())
-        })?;
-    }
-
-    Ok(paths.managed_clickhouse_dir.display().to_string())
-}
-
-fn cmd_config_get(cfg: &AppConfig, key: &str) -> Result<String> {
-    match key {
-        "clickhouse.url" => Ok(cfg.clickhouse.url.clone()),
-        "clickhouse.database" => Ok(cfg.clickhouse.database.clone()),
-        _ => bail!(
-            "unsupported config key '{}'; supported keys: clickhouse.url, clickhouse.database",
-            key
-        ),
-    }
-}
-
-fn health_label(value: bool) -> &'static str {
-    if value {
-        "healthy"
-    } else {
-        "unhealthy"
-    }
-}
-
-fn state_label(value: bool) -> &'static str {
-    if value {
-        "yes"
-    } else {
-        "no"
-    }
-}
-
-fn stoplight(running: bool) -> &'static str {
-    if running {
-        "\u{1F7E2}" // 🟢
-    } else {
-        "\u{1F534}" // 🔴
-    }
-}
-
-fn service_endpoint(service: Service, snapshot: &StatusSnapshot) -> Option<String> {
-    match service {
-        Service::ClickHouse => Some(snapshot.clickhouse_health_url.clone()),
-        Service::Monitor => snapshot.monitor_url.clone(),
-        _ => None,
-    }
-}
-
-fn format_start_state(outcome: &StartOutcome) -> String {
-    match outcome.state {
-        StartState::Started => "started".to_string(),
-        StartState::AlreadyRunning => "already running".to_string(),
-        StartState::AlreadyServing => "already serving (unmanaged)".to_string(),
-    }
-}
-
-fn format_start_pid(outcome: &StartOutcome) -> String {
-    match outcome.pid {
-        Some(pid) => pid.to_string(),
-        None => "-".to_string(),
-    }
-}
-
-fn render_status(output: &CliOutput, snapshot: &StatusSnapshot) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(snapshot)?);
-        return Ok(());
-    }
-
-    // -- Services with stoplight indicators and endpoints --
-    let service_rows: Vec<Vec<String>> = snapshot
-        .services
-        .iter()
-        .map(|row| {
-            let running = row.pid.is_some();
-            let mut cols = vec![
-                format!("{} {}", stoplight(running), row.service.name()),
-                if running {
-                    "running".to_string()
-                } else {
-                    "stopped".to_string()
-                },
-                service_endpoint(row.service, snapshot).unwrap_or_default(),
-            ];
-            if output.verbose {
-                cols.push(
-                    row.pid
-                        .map(|pid| pid.to_string())
-                        .unwrap_or_else(|| "-".to_string()),
-                );
-            }
-            cols
-        })
-        .collect();
-
-    if output.verbose {
-        output.table("Services", &["", "state", "endpoint", "pid"], &service_rows);
-    } else {
-        output.table("Services", &["", "state", "endpoint"], &service_rows);
-    }
-
-    // -- Database Health (concise) --
-    let db_healthy = snapshot.doctor.clickhouse_healthy && snapshot.doctor.database_exists;
-    let mut doctor_lines = vec![format!(
-        "{} {}",
-        stoplight(db_healthy),
-        if db_healthy {
-            "database healthy".to_string()
-        } else {
-            format!(
-                "clickhouse {} / db {}",
-                health_label(snapshot.doctor.clickhouse_healthy),
-                if snapshot.doctor.database_exists {
-                    "exists"
-                } else {
-                    "missing"
-                }
-            )
-        }
-    )];
-    if let Some(version) = &snapshot.doctor.clickhouse_version {
-        doctor_lines[0].push_str(&format!("  (v{version})"));
-    }
-    if !snapshot.doctor.pending_migrations.is_empty() {
-        doctor_lines.push(format!(
-            "  pending migrations: {}",
-            snapshot.doctor.pending_migrations.join(", ")
-        ));
-    }
-    if !snapshot.doctor.missing_tables.is_empty() {
-        doctor_lines.push(format!(
-            "  missing tables: {}",
-            snapshot.doctor.missing_tables.join(", ")
-        ));
-    }
-    if output.verbose && !snapshot.doctor.errors.is_empty() {
-        doctor_lines.push(format!("  errors: {}", snapshot.doctor.errors.join(" | ")));
-    }
-    output.section("Database", &doctor_lines);
-
-    // -- Ingest activity (only show when there is something to report) --
-    match &snapshot.heartbeat {
-        HeartbeatSnapshot::Available {
-            latest,
-            queue_depth,
-            files_active,
-            watcher_backend,
-            watcher_error_count,
-            watcher_reset_count,
-            watcher_last_reset_unix_ms,
-        } => {
-            let mut lines = vec![
-                format!("last event: {latest}"),
-                format!("queue: {queue_depth}  |  active files: {files_active}"),
-            ];
-            if *watcher_error_count > 0 || *watcher_reset_count > 0 {
-                lines.push(format!(
-                    "watcher: {watcher_backend}  (errors: {watcher_error_count}, resets: {watcher_reset_count})"
-                ));
-            } else if output.verbose {
-                lines.push(format!("watcher: {watcher_backend}"));
-            }
-            if output.verbose {
-                lines.push(format!(
-                    "watcher last reset unix ms: {watcher_last_reset_unix_ms}"
-                ));
-            }
-            output.section("Ingest", &lines);
-        }
-        HeartbeatSnapshot::Unavailable => {
-            if output.verbose {
-                output.section("Ingest", &["no heartbeat data".to_string()]);
-            }
-        }
-        HeartbeatSnapshot::Error { message } => {
-            output.section("Ingest", &[format!("heartbeat error: {message}")]);
-        }
-    }
-
-    // -- ClickHouse runtime details (verbose only) --
-    if output.verbose {
-        let mut ch_lines = vec![
-            format!(
-                "managed install: {}",
-                if snapshot.managed_clickhouse_installed {
-                    "present"
-                } else {
-                    "missing"
-                }
-            ),
-            format!("binary: {}", snapshot.managed_clickhouse_path),
-            format!(
-                "source: {}{}",
-                snapshot.clickhouse_active_source,
-                snapshot
-                    .clickhouse_active_source_path
-                    .as_ref()
-                    .map(|p| format!(" ({p})"))
-                    .unwrap_or_default()
-            ),
-            format!("checksum: {}", snapshot.managed_clickhouse_checksum),
-        ];
-        if let Some(version) = &snapshot.managed_clickhouse_version {
-            ch_lines.push(format!("managed version: {version}"));
-        }
-        output.section("ClickHouse Runtime", &ch_lines);
-    }
-
-    // -- Status notes (warnings) --
-    if !snapshot.status_notes.is_empty() {
-        output.section("Warnings", &snapshot.status_notes);
-    }
-    Ok(())
-}
-
-fn render_db_migrate(output: &CliOutput, outcome: &MigrationOutcome) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(outcome)?);
-        return Ok(());
-    }
-    if outcome.applied.is_empty() {
-        output.section("Database Migrations", &["already up to date".to_string()]);
-        return Ok(());
-    }
-    let rows = outcome
-        .applied
-        .iter()
-        .enumerate()
-        .map(|(idx, migration)| vec![(idx + 1).to_string(), migration.to_string()])
-        .collect::<Vec<_>>();
-    output.table("Applied Migrations", &["#", "migration"], &rows);
-    Ok(())
-}
-
-fn doctor_is_healthy(report: &DoctorReport) -> bool {
-    report.clickhouse_healthy
-        && report.database_exists
-        && report.pending_migrations.is_empty()
-        && report.missing_tables.is_empty()
-        && report.errors.is_empty()
-}
-
-fn render_db_doctor(output: &CliOutput, report: &DoctorReport) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(report)?);
-        return Ok(());
-    }
-
-    let mut lines = vec![
-        format!("clickhouse: {}", health_label(report.clickhouse_healthy)),
-        format!("database: {}", report.database),
-        format!("database exists: {}", state_label(report.database_exists)),
-        format!(
-            "pending migrations: {}",
-            if report.pending_migrations.is_empty() {
-                "none".to_string()
-            } else {
-                report.pending_migrations.join(", ")
-            }
-        ),
-        format!(
-            "missing tables: {}",
-            if report.missing_tables.is_empty() {
-                "none".to_string()
-            } else {
-                report.missing_tables.join(", ")
-            }
-        ),
-    ];
-    if let Some(version) = &report.clickhouse_version {
-        lines.push(format!("clickhouse version: {version}"));
-    }
-    if output.verbose && !report.applied_migrations.is_empty() {
-        lines.push(format!(
-            "applied migrations: {}",
-            report.applied_migrations.join(", ")
-        ));
-    }
-    if !report.errors.is_empty() {
-        lines.push(format!("errors: {}", report.errors.join(" | ")));
-    }
-    output.section("DB Doctor", &lines);
-    Ok(())
-}
-
-fn render_logs(output: &CliOutput, snapshot: &LogsSnapshot) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(snapshot)?);
-        return Ok(());
-    }
-    for section in &snapshot.sections {
-        let mut lines = vec![
-            format!("path: {}", section.path),
-            format!("lines requested: {}", snapshot.requested_lines),
-        ];
-        if !section.exists {
-            lines.push("log file: missing".to_string());
-            output.section(&format!("Logs: {}", section.service.name()), &lines);
-            continue;
-        }
-        lines.push(format!("lines returned: {}", section.lines.len()));
-        output.section(&format!("Logs: {}", section.service.name()), &lines);
-        for line in &section.lines {
-            output.line(line);
-        }
-    }
-    Ok(())
-}
-
-fn render_clickhouse_status(output: &CliOutput, snapshot: &ClickhouseStatusSnapshot) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(snapshot)?);
-        return Ok(());
-    }
-    let mut lines = vec![
-        format!("managed root: {}", snapshot.managed_root),
-        format!(
-            "clickhouse binary: {}",
-            state_label(snapshot.clickhouse_exists)
-        ),
-        format!(
-            "clickhouse-server binary: {}",
-            state_label(snapshot.clickhouse_server_exists)
-        ),
-        format!(
-            "clickhouse-client binary: {}",
-            state_label(snapshot.clickhouse_client_exists)
-        ),
-        format!("expected version: {}", snapshot.expected_version),
-        format!(
-            "active source: {}{}",
-            snapshot.active_source,
-            snapshot
-                .active_source_path
-                .as_ref()
-                .map(|p| format!(" ({p})"))
-                .unwrap_or_default()
-        ),
-        format!("checksum state: {}", snapshot.checksum_state),
-    ];
-    if let Some(version) = &snapshot.installed_version {
-        lines.push(format!("installed version: {version}"));
-    }
-    output.section("Managed ClickHouse", &lines);
-    Ok(())
-}
-
-fn render_up(output: &CliOutput, snapshot: &UpSnapshot) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(snapshot)?);
-        return Ok(());
-    }
-    let mut rows = vec![vec![
-        snapshot.clickhouse.service.name().to_string(),
-        format_start_state(&snapshot.clickhouse),
-        format_start_pid(&snapshot.clickhouse),
-    ]];
-    rows.extend(snapshot.services.iter().map(|outcome| {
-        vec![
-            outcome.service.name().to_string(),
-            format_start_state(outcome),
-            format_start_pid(outcome),
-        ]
-    }));
-    output.table("Startup Results", &["service", "result", "pid"], &rows);
-    render_db_migrate(output, &snapshot.migrations)?;
-    render_status(output, &snapshot.status)?;
-    Ok(())
-}
-
-fn render_down(output: &CliOutput, snapshot: &DownSnapshot) -> Result<()> {
-    if output.is_json() {
-        println!("{}", serde_json::to_string_pretty(snapshot)?);
-        return Ok(());
-    }
-    if snapshot.stopped.is_empty() {
-        output.section("Shutdown", &["no running services found".to_string()]);
-        if let Some(warning) = &snapshot.warning {
-            output.section("Warning", std::slice::from_ref(warning));
-        }
-        return Ok(());
-    }
-    let rows = snapshot
-        .stopped
-        .iter()
-        .map(|service| vec![service.name().to_string(), "stopped".to_string()])
-        .collect::<Vec<_>>();
-    output.table("Shutdown", &["service", "result"], &rows);
-    if let Some(warning) = &snapshot.warning {
-        output.section("Warning", std::slice::from_ref(warning));
-    }
-    Ok(())
-}
+mod cli;
+mod commands;
+mod managed_clickhouse;
+mod paths;
+mod process;
+mod render;
+mod service;
+
+use anyhow::Result;
+use clap::Parser;
+use std::process::ExitCode;
+
+use crate::cli::Cli;
+use crate::render::CliOutput;
 
 #[tokio::main(flavor = "multi_thread")]
 async fn main() -> Result<ExitCode> {
     let cli = Cli::parse();
     let output = CliOutput::from_cli(&cli);
-
-    match cli.command {
-        CliCommand::Up(args) => {
-            let (config_path, cfg) = load_cfg(cli.config.clone())?;
-            let paths = runtime_paths(&cfg);
-            let services_to_start = selected_up_services(&args, &cfg);
-            preflight_required_service_binaries(&services_to_start, &paths)?;
-            ensure_runtime_dirs(&paths)?;
-
-            let clickhouse = start_clickhouse(&cfg, &paths).await?;
-            let migrations = cmd_db_migrate(&cfg).await?;
-
-            let mut started_services = Vec::new();
-            for service in services_to_start {
-                // The up-managed Mcp service IS the shared central server;
-                // there is no up-managed stdio variant. The `--serve socket`
-                // flag is injected here (NOT in the shared
-                // service_args_with_defaults helper) so the foreground
-                // `moraine run mcp` path that agents register stays a stdio
-                // client and never accidentally becomes a server. The explicit
-                // `--socket` pins the daemon to the launcher-resolved path so it
-                // binds exactly where clients (which resolve from the same
-                // config) connect.
-                let extra_args: Vec<String> = if service == Service::Mcp {
-                    vec![
-                        "--serve".to_string(),
-                        "socket".to_string(),
-                        "--socket".to_string(),
-                        cfg.mcp.central_socket_path.clone(),
-                    ]
-                } else {
-                    Vec::new()
-                };
-                started_services.push(start_background_service(
-                    service,
-                    &config_path,
-                    &cfg,
-                    &paths,
-                    &extra_args,
-                )?);
-            }
-
-            let status = cmd_status(&paths, &cfg).await?;
-            let snapshot = UpSnapshot {
-                clickhouse,
-                migrations,
-                services: started_services,
-                status,
-            };
-            render_up(&output, &snapshot)?;
-            Ok(ExitCode::SUCCESS)
-        }
-        CliCommand::Down => {
-            let (_, cfg) = load_cfg(cli.config.clone())?;
-            let paths = runtime_paths(&cfg);
-            let mut stopped = Vec::new();
-            for service in [
-                Service::Mcp,
-                Service::Monitor,
-                Service::Ingest,
-                Service::ClickHouse,
-            ] {
-                if stop_service(&paths, service)? {
-                    stopped.push(service);
-                }
-            }
-            // Remove the central MCP socket only when nothing is listening on
-            // it: a dead socket would make clients connect-then-stall before
-            // falling back to embedded, but unlinking a *live* server's socket
-            // (possible when its PID file was lost, so stop_service had nothing
-            // to kill) would orphan that server unreachably while it keeps
-            // running. A daemon stopped by stop_service above has already
-            // removed its own socket via its signal handler.
-            let warning = if central_socket_live(&cfg.mcp.central_socket_path) {
-                Some(format!(
-                    "an MCP server is still listening on {} but is not tracked by a \
-                     PID file, so it was not stopped and its socket was left in place; \
-                     stop that process manually and re-run `moraine down`",
-                    cfg.mcp.central_socket_path
-                ))
-            } else {
-                let _ = fs::remove_file(&cfg.mcp.central_socket_path);
-                None
-            };
-            render_down(&output, &DownSnapshot { stopped, warning })?;
-            Ok(ExitCode::SUCCESS)
-        }
-        CliCommand::Status => {
-            let (_, cfg) = load_cfg(cli.config.clone())?;
-            let paths = runtime_paths(&cfg);
-            let snapshot = cmd_status(&paths, &cfg).await?;
-            render_status(&output, &snapshot)?;
-            Ok(ExitCode::SUCCESS)
-        }
-        CliCommand::Logs(args) => {
-            let (_, cfg) = load_cfg(cli.config.clone())?;
-            let paths = runtime_paths(&cfg);
-            let snapshot = collect_logs(&paths, args.service, args.lines)?;
-            render_logs(&output, &snapshot)?;
-            Ok(ExitCode::SUCCESS)
-        }
-        CliCommand::Db(args) => {
-            let (_, cfg) = load_cfg(cli.config.clone())?;
-            match args.command {
-                DbCommand::Migrate => {
-                    let outcome = cmd_db_migrate(&cfg).await?;
-                    render_db_migrate(&output, &outcome)?;
-                    Ok(ExitCode::SUCCESS)
-                }
-                DbCommand::Doctor => {
-                    let report = cmd_db_doctor(&cfg).await?;
-                    render_db_doctor(&output, &report)?;
-                    if doctor_is_healthy(&report) {
-                        Ok(ExitCode::SUCCESS)
-                    } else {
-                        Ok(ExitCode::from(1))
-                    }
-                }
-            }
-        }
-        CliCommand::Clickhouse(args) => {
-            let (_, cfg) = load_cfg(cli.config.clone())?;
-            let paths = runtime_paths(&cfg);
-            match args.command {
-                ClickhouseCommand::Install(install) => {
-                    let version = install
-                        .version
-                        .unwrap_or_else(|| cfg.runtime.clickhouse_version.clone());
-                    let installed = cmd_clickhouse_install(&paths, &version, install.force).await?;
-                    if output.is_json() {
-                        println!(
-                            "{}",
-                            serde_json::to_string_pretty(&serde_json::json!({
-                                "installed_path": installed.display().to_string(),
-                                "version": version,
-                                "force": install.force,
-                            }))?
-                        );
-                    } else {
-                        output.section(
-                            "Managed ClickHouse Install",
-                            &[
-                                format!("installed binary: {}", installed.display()),
-                                format!("version: {version}"),
-                                format!("force: {}", state_label(install.force)),
-                            ],
-                        );
-                    }
-                    Ok(ExitCode::SUCCESS)
-                }
-                ClickhouseCommand::Status => {
-                    let snapshot = cmd_clickhouse_status(&cfg, &paths);
-                    render_clickhouse_status(&output, &snapshot)?;
-                    Ok(ExitCode::SUCCESS)
-                }
-                ClickhouseCommand::Uninstall => {
-                    let removed = cmd_clickhouse_uninstall(&paths)?;
-                    if output.is_json() {
-                        println!(
-                            "{}",
-                            serde_json::to_string_pretty(&serde_json::json!({
-                                "removed_path": removed
-                            }))?
-                        );
-                    } else {
-                        output.section(
-                            "Managed ClickHouse Uninstall",
-                            &[format!("removed: {removed}")],
-                        );
-                    }
-                    Ok(ExitCode::SUCCESS)
-                }
-            }
-        }
-        CliCommand::Config(args) => {
-            let (_, cfg) = load_cfg(cli.config.clone())?;
-            match args.command {
-                ConfigCommand::Get(get) => {
-                    let value = cmd_config_get(&cfg, &get.key)?;
-                    if output.is_json() {
-                        println!(
-                            "{}",
-                            serde_json::to_string_pretty(&serde_json::json!({
-                                "key": get.key,
-                                "value": value,
-                            }))?
-                        );
-                    } else {
-                        println!("{value}");
-                    }
-                    Ok(ExitCode::SUCCESS)
-                }
-            }
-        }
-        CliCommand::Run(run) => {
-            let (inline_config, passthrough) = parse_config_flag(&run.args)?;
-            let raw_config = inline_config.or(cli.config.clone());
-            let (config_path, cfg) = load_cfg(raw_config)?;
-            let paths = runtime_paths(&cfg);
-            run_foreground_service(run.service, &config_path, &cfg, &paths, &passthrough).await
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::ffi::OsString;
-    use std::sync::{Mutex, MutexGuard};
-
-    static ENV_VAR_LOCK: Mutex<()> = Mutex::new(());
-
-    struct EnvVarGuard {
-        key: &'static str,
-        original: Option<OsString>,
-    }
-
-    impl EnvVarGuard {
-        fn capture(key: &'static str) -> Self {
-            Self {
-                key,
-                original: std::env::var_os(key),
-            }
-        }
-    }
-
-    impl Drop for EnvVarGuard {
-        fn drop(&mut self) {
-            if let Some(value) = &self.original {
-                std::env::set_var(self.key, value);
-            } else {
-                std::env::remove_var(self.key);
-            }
-        }
-    }
-
-    fn lock_env_vars() -> MutexGuard<'static, ()> {
-        ENV_VAR_LOCK.lock().expect("env-var lock poisoned")
-    }
-
-    fn temp_dir(name: &str) -> PathBuf {
-        let stamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_nanos();
-        let path = std::env::temp_dir().join(format!("moraine-{name}-{stamp}"));
-        fs::create_dir_all(&path).expect("create temp dir");
-        path
-    }
-
-    fn write_file(path: &Path) {
-        fs::create_dir_all(path.parent().expect("parent")).expect("create parent");
-        fs::write(path, b"#!/bin/sh\n").expect("write file");
-    }
-
-    fn test_doctor_report(clickhouse_healthy: bool) -> DoctorReport {
-        DoctorReport {
-            clickhouse_healthy,
-            clickhouse_version: None,
-            database: "moraine".to_string(),
-            database_exists: true,
-            applied_migrations: Vec::new(),
-            pending_migrations: Vec::new(),
-            missing_tables: Vec::new(),
-            errors: Vec::new(),
-        }
-    }
-
-    #[test]
-    fn tail_lines_returns_last_n_without_trailing_newline() {
-        let root = temp_dir("tail-lines-basic");
-        let path = root.join("test.log");
-        fs::write(&path, "one\ntwo\nthree").expect("write log");
-
-        let lines = tail_lines(&path, 2).expect("tail lines");
-        assert_eq!(lines, vec!["two".to_string(), "three".to_string()]);
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn tail_lines_handles_utf8_chunk_boundary() {
-        let root = temp_dir("tail-lines-utf8");
-        let path = root.join("test.log");
-        let prefix = "é".repeat(4500);
-        let content = format!("{prefix}\nmiddle\ntail\n");
-        fs::write(&path, content).expect("write log");
-
-        let one = tail_lines(&path, 1).expect("tail one line");
-        assert_eq!(one, vec!["tail".to_string()]);
-
-        let two = tail_lines(&path, 2).expect("tail two lines");
-        assert_eq!(two, vec!["middle".to_string(), "tail".to_string()]);
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn clickhouse_asset_selection_covers_target_matrix() {
-        let linux_x64 =
-            clickhouse_asset_for_target(DEFAULT_CLICKHOUSE_TAG, "x86_64-unknown-linux-gnu")
-                .expect("linux x64");
-        assert_eq!(linux_x64.url, CH_URL_LINUX_X86_64);
-        assert!(linux_x64.is_archive);
-
-        let linux_arm =
-            clickhouse_asset_for_target(DEFAULT_CLICKHOUSE_TAG, "aarch64-unknown-linux-gnu")
-                .expect("linux arm");
-        assert_eq!(linux_arm.url, CH_URL_LINUX_AARCH64);
-        assert!(linux_arm.is_archive);
-
-        let mac_x64 = clickhouse_asset_for_target(DEFAULT_CLICKHOUSE_TAG, "x86_64-apple-darwin")
-            .expect("mac x64");
-        assert_eq!(mac_x64.url, CH_URL_MACOS_X86_64);
-        assert!(!mac_x64.is_archive);
-
-        let mac_arm = clickhouse_asset_for_target(DEFAULT_CLICKHOUSE_TAG, "aarch64-apple-darwin")
-            .expect("mac arm");
-        assert_eq!(mac_arm.url, CH_URL_MACOS_AARCH64);
-        assert!(!mac_arm.is_archive);
-    }
-
-    #[test]
-    fn clickhouse_asset_rejects_unsupported_version() {
-        let err = clickhouse_asset_for_target("v0.0.0", "x86_64-unknown-linux-gnu")
-            .expect_err("unsupported version");
-        assert!(
-            err.to_string().contains("unsupported ClickHouse version"),
-            "{}",
-            err
-        );
-    }
-
-    #[test]
-    fn clickhouse_logs_use_internal_rotating_path() {
-        let root = temp_dir("clickhouse-log-path");
-        let logs_dir = root.join("logs");
-
-        let mut cfg = AppConfig::default();
-        cfg.runtime.root_dir = root.to_string_lossy().to_string();
-        cfg.runtime.logs_dir = logs_dir.to_string_lossy().to_string();
-        let paths = runtime_paths(&cfg);
-
-        assert_eq!(
-            log_path(&paths, Service::ClickHouse),
-            root.join("clickhouse/log/clickhouse-server.log")
-        );
-        assert_eq!(
-            log_path(&paths, Service::Ingest),
-            logs_dir.join("ingest.log")
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn cleanup_legacy_clickhouse_pipe_log_removes_legacy_file() {
-        let root = temp_dir("legacy-clickhouse-log");
-        let logs_dir = root.join("logs");
-        fs::create_dir_all(&logs_dir).expect("create logs dir");
-
-        let mut cfg = AppConfig::default();
-        cfg.runtime.root_dir = root.to_string_lossy().to_string();
-        cfg.runtime.logs_dir = logs_dir.to_string_lossy().to_string();
-        let paths = runtime_paths(&cfg);
-
-        let legacy_log = legacy_clickhouse_pipe_log_path(&paths);
-        fs::write(&legacy_log, b"legacy clickhouse stdout").expect("write legacy log");
-        assert!(legacy_log.exists());
-
-        cleanup_legacy_clickhouse_pipe_log(&paths);
-        assert!(!legacy_log.exists());
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn find_file_ending_with_prefers_usr_bin_clickhouse() {
-        let root = temp_dir("find-file-ending-with");
-        let completion = root.join("pkg/usr/share/bash-completion/completions/clickhouse");
-        let binary = root.join("pkg/usr/bin/clickhouse");
-        write_file(&completion);
-        write_file(&binary);
-
-        let resolved = find_file_ending_with(&root, &["usr", "bin", "clickhouse"])
-            .expect("resolve clickhouse path")
-            .expect("clickhouse path");
-
-        assert_eq!(resolved, binary);
-    }
-
-    #[test]
-    fn build_status_notes_flags_healthy_external_clickhouse() {
-        let services = vec![ServiceRuntimeStatus {
-            service: Service::ClickHouse,
-            pid: None,
-        }];
-        let report = test_doctor_report(true);
-        let notes = build_status_notes(&services, &report, "http://127.0.0.1:8123");
-        assert_eq!(notes.len(), 1);
-        assert!(
-            notes[0].contains("endpoint is healthy while managed clickhouse runtime is stopped")
-        );
-        assert!(notes[0].contains("http://127.0.0.1:8123"));
-    }
-
-    #[test]
-    fn build_status_notes_flags_unhealthy_managed_clickhouse() {
-        let services = vec![ServiceRuntimeStatus {
-            service: Service::ClickHouse,
-            pid: Some(4242),
-        }];
-        let report = test_doctor_report(false);
-        let notes = build_status_notes(&services, &report, "http://127.0.0.1:8123");
-        assert_eq!(notes.len(), 1);
-        assert!(notes[0].contains("managed clickhouse runtime is running"));
-        assert!(notes[0].contains("are failing"));
-        assert!(notes[0].contains("http://127.0.0.1:8123"));
-    }
-
-    #[test]
-    fn clap_parses_clickhouse_install_flags() {
-        let cli = Cli::parse_from([
-            "moraine",
-            "clickhouse",
-            "install",
-            "--version",
-            "v25.12.5.44-stable",
-            "--force",
-        ]);
-        match cli.command {
-            CliCommand::Clickhouse(ClickhouseArgs {
-                command: ClickhouseCommand::Install(install),
-            }) => {
-                assert!(install.force);
-                assert_eq!(install.version.as_deref(), Some("v25.12.5.44-stable"));
-            }
-            _ => panic!("expected clickhouse install command"),
-        }
-    }
-
-    #[test]
-    fn clap_parses_config_get_key() {
-        let cli = Cli::parse_from(["moraine", "config", "get", "clickhouse.url"]);
-        match cli.command {
-            CliCommand::Config(ConfigArgs {
-                command: ConfigCommand::Get(get),
-            }) => assert_eq!(get.key, "clickhouse.url"),
-            _ => panic!("expected config get command"),
-        }
-    }
-
-    #[test]
-    fn clap_parses_run_passthrough_args() {
-        let cli = Cli::parse_from([
-            "moraine",
-            "--output",
-            "plain",
-            "run",
-            "mcp",
-            "--",
-            "--stdio",
-            "--transport",
-            "jsonrpc",
-        ]);
-        match cli.command {
-            CliCommand::Run(run) => {
-                assert_eq!(run.service, Service::Mcp);
-                assert_eq!(
-                    run.args,
-                    vec![
-                        "--stdio".to_string(),
-                        "--transport".to_string(),
-                        "jsonrpc".to_string(),
-                    ]
-                );
-            }
-            _ => panic!("expected run command"),
-        }
-    }
-
-    #[test]
-    fn cmd_config_get_returns_supported_keys() {
-        let mut cfg = AppConfig::default();
-        cfg.clickhouse.url = "http://127.0.0.1:18123".to_string();
-        cfg.clickhouse.database = "analytics".to_string();
-
-        assert_eq!(
-            cmd_config_get(&cfg, "clickhouse.url").expect("url"),
-            "http://127.0.0.1:18123"
-        );
-        assert_eq!(
-            cmd_config_get(&cfg, "clickhouse.database").expect("database"),
-            "analytics"
-        );
-    }
-
-    #[test]
-    fn cmd_config_get_rejects_unknown_key() {
-        let cfg = AppConfig::default();
-        let err = cmd_config_get(&cfg, "runtime.root_dir").expect_err("unknown key");
-        assert!(err.to_string().contains("unsupported config key"));
-    }
-
-    #[test]
-    fn output_mode_respects_json_flag() {
-        let cli = Cli::parse_from(["moraine", "--output", "json", "status"]);
-        let output = CliOutput::from_cli(&cli);
-        assert_eq!(output.mode, OutputMode::Json);
-    }
-
-    #[test]
-    fn clickhouse_ports_follow_url_port() {
-        let mut cfg = AppConfig::default();
-        cfg.clickhouse.url = "http://127.0.0.1:18123".to_string();
-        let ports = clickhouse_ports_from_url(&cfg).expect("ports");
-        assert_eq!(ports, (18123, 19000, 19009));
-    }
-
-    #[test]
-    fn clickhouse_ports_require_valid_url() {
-        let mut cfg = AppConfig::default();
-        cfg.clickhouse.url = "not-a-url".to_string();
-        let err = clickhouse_ports_from_url(&cfg).expect_err("invalid url");
-        assert!(err.to_string().contains("invalid clickhouse.url"));
-    }
-
-    #[test]
-    fn clickhouse_url_is_local_accepts_loopback_hosts() {
-        for url in [
-            "http://localhost:8123",
-            "http://127.0.0.1:8123",
-            "http://[::1]:8123",
-            "http://0.0.0.0:8123",
-            "http://[::]:8123",
-        ] {
-            let mut cfg = AppConfig::default();
-            cfg.clickhouse.url = url.to_string();
-            assert!(clickhouse_url_is_local(&cfg).expect(url), "{url}");
-        }
-    }
-
-    #[test]
-    fn clickhouse_url_is_local_rejects_remote_hosts() {
-        for url in [
-            "http://clickhouse.local:8123",
-            "https://clickhouse.example.com:8443",
-            "http://192.168.1.50:8123",
-        ] {
-            let mut cfg = AppConfig::default();
-            cfg.clickhouse.url = url.to_string();
-            assert!(!clickhouse_url_is_local(&cfg).expect(url), "{url}");
-        }
-    }
-
-    #[test]
-    fn monitor_runtime_url_uses_configured_bind() {
-        let mut cfg = AppConfig::default();
-        cfg.monitor.host = "127.0.0.1".to_string();
-        cfg.monitor.port = 18080;
-        assert_eq!(monitor_runtime_url(&cfg), "http://127.0.0.1:18080");
-    }
-
-    #[test]
-    fn monitor_runtime_url_wraps_ipv6_host() {
-        let mut cfg = AppConfig::default();
-        cfg.monitor.host = "::1".to_string();
-        cfg.monitor.port = 18080;
-        assert_eq!(monitor_runtime_url(&cfg), "http://[::1]:18080");
-    }
-
-    #[test]
-    fn monitor_runtime_running_checks_monitor_pid() {
-        let services = vec![
-            ServiceRuntimeStatus {
-                service: Service::ClickHouse,
-                pid: Some(100),
-            },
-            ServiceRuntimeStatus {
-                service: Service::Monitor,
-                pid: Some(200),
-            },
-        ];
-        assert!(monitor_runtime_running(&services));
-
-        let stopped_monitor = vec![ServiceRuntimeStatus {
-            service: Service::Monitor,
-            pid: None,
-        }];
-        assert!(!monitor_runtime_running(&stopped_monitor));
-    }
-
-    #[test]
-    fn resolve_service_binary_prefers_env_then_config() {
-        let _env_lock = lock_env_vars();
-        let _service_bin_dir_guard = EnvVarGuard::capture("MORAINE_SERVICE_BIN_DIR");
-        let _source_tree_mode_guard = EnvVarGuard::capture("MORAINE_SOURCE_TREE_MODE");
-
-        let root = temp_dir("resolver");
-        let env_dir = root.join("env");
-        let cfg_dir = root.join("cfg");
-        let env_bin = env_dir.join("moraine-ingest");
-        let cfg_bin = cfg_dir.join("moraine-ingest");
-        write_file(&env_bin);
-        write_file(&cfg_bin);
-
-        let mut cfg = AppConfig::default();
-        cfg.runtime.service_bin_dir = cfg_dir.to_string_lossy().to_string();
-        let paths = runtime_paths(&cfg);
-
-        std::env::remove_var("MORAINE_SOURCE_TREE_MODE");
-        std::env::set_var("MORAINE_SERVICE_BIN_DIR", &env_dir);
-        assert_eq!(
-            resolve_service_binary(Service::Ingest, &paths).resolved_path,
-            Some(env_bin.clone())
-        );
-
-        std::env::remove_var("MORAINE_SERVICE_BIN_DIR");
-        assert_eq!(
-            resolve_service_binary(Service::Ingest, &paths).resolved_path,
-            Some(cfg_bin)
-        );
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn resolve_service_binary_reports_missing_without_path_fallback() {
-        let _env_lock = lock_env_vars();
-        let _service_bin_dir_guard = EnvVarGuard::capture("MORAINE_SERVICE_BIN_DIR");
-        let _source_tree_mode_guard = EnvVarGuard::capture("MORAINE_SOURCE_TREE_MODE");
-
-        let root = temp_dir("resolver-path");
-        let mut cfg = AppConfig::default();
-        cfg.runtime.service_bin_dir = root.join("missing").to_string_lossy().to_string();
-        let paths = runtime_paths(&cfg);
-
-        std::env::remove_var("MORAINE_SERVICE_BIN_DIR");
-        std::env::remove_var("MORAINE_SOURCE_TREE_MODE");
-        let resolved = resolve_service_binary(Service::Mcp, &paths);
-        assert_eq!(resolved.binary_name, "moraine-mcp");
-        assert!(resolved.resolved_path.is_none());
-        assert!(resolved
-            .checked_paths
-            .iter()
-            .any(|probe| probe.source == "runtime.service_bin_dir"
-                && probe.path == paths.service_bin_dir.join("moraine-mcp")));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn require_service_binary_includes_remediation() {
-        let _env_lock = lock_env_vars();
-        let _service_bin_dir_guard = EnvVarGuard::capture("MORAINE_SERVICE_BIN_DIR");
-        let _source_tree_mode_guard = EnvVarGuard::capture("MORAINE_SOURCE_TREE_MODE");
-
-        let root = temp_dir("resolver-remediation");
-        let mut cfg = AppConfig::default();
-        cfg.runtime.service_bin_dir = root.join("missing").to_string_lossy().to_string();
-        let paths = runtime_paths(&cfg);
-
-        std::env::remove_var("MORAINE_SERVICE_BIN_DIR");
-        std::env::remove_var("MORAINE_SOURCE_TREE_MODE");
-        let err = require_service_binary(Service::Mcp, &paths).expect_err("missing mcp binary");
-        let message = err.to_string();
-        assert!(message.contains("required service binary `moraine-mcp`"));
-        assert!(message.contains("runtime.service_bin_dir"));
-        assert!(message.contains("MORAINE_SERVICE_BIN_DIR"));
-        assert!(message.contains("cargo build --workspace --locked"));
-        assert!(message.contains("does not fall back to PATH"));
-
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn managed_checksum_state_reports_verified() {
-        let root = temp_dir("checksum");
-        let managed_dir = root.join("managed");
-        fs::create_dir_all(&managed_dir).expect("managed dir");
-
-        let mut cfg = AppConfig::default();
-        cfg.runtime.managed_clickhouse_dir = managed_dir.to_string_lossy().to_string();
-        let paths = runtime_paths(&cfg);
-
-        let expected = clickhouse_asset_for_host(&cfg.runtime.clickhouse_version)
-            .expect("host asset")
-            .sha256;
-        fs::write(
-            managed_clickhouse_checksum_file(&paths),
-            format!("{expected}\n"),
-        )
-        .expect("write checksum");
-
-        assert_eq!(managed_clickhouse_checksum_state(&cfg, &paths), "verified");
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[cfg(unix)]
-    #[test]
-    fn central_socket_live_detects_listener_vs_dead_socket() {
-        // Keep the path short: macOS temp_dir() exceeds the ~104-byte sun_path
-        // limit for Unix socket addresses.
-        let stamp = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .unwrap_or_default()
-            .as_nanos();
-        let sock = PathBuf::from(format!(
-            "/tmp/moraine-csl-{}-{stamp}.sock",
-            std::process::id()
-        ));
-        let sock_str = sock.to_string_lossy().to_string();
-
-        // No socket file at all.
-        assert!(!central_socket_live(&sock_str));
-
-        // A live listener is detected even with no PID file anywhere.
-        let listener = std::os::unix::net::UnixListener::bind(&sock).expect("bind socket");
-        assert!(central_socket_live(&sock_str));
-
-        // A dead socket file (listener gone) is not treated as live.
-        drop(listener);
-        assert!(!central_socket_live(&sock_str));
-
-        let _ = fs::remove_file(&sock);
-    }
+    commands::dispatch(cli, output).await
 }
